[Python] Add Python package namespace
- Match src dir with package name
- Update thrift gen commands
