+++ lib/python/cmssdk/AdminAuth/__init__.py
