+++ lib/python/cmssdk/AdminAuth/constants.py
