+++ lib/python/cmssdk/AdminMenu/AdminMenuService.py
@@ -18,61 +18,46 @@ from thrift.transport import TTransport
 
 class Iface(object):
     """
-    AdminUser 서비스
+    AdminMenu 서비스
     """
-    def getAllAdminUserArray(self):
-        pass
-
-    def getUser(self, userId):
+    def getMenuList(self, isUse):
         """
         Parameters:
-         - userId
-        """
-        pass
-
-    def getAdminUserTag(self, userId):
-        """
-        Parameters:
-         - userId
-        """
-        pass
-
-    def getAdminUserMenu(self, userId):
-        """
-        Parameters:
-         - userId
-        """
-        pass
-
-    def getAllMenuIds(self, userId):
-        """
-        Parameters:
-         - userId
-        """
-        pass
-
-    def updateMyInfo(self, name, team, isUse, passwd):
-        """
-        Parameters:
-         - name
-         - team
          - isUse
-         - passwd
         """
         pass
 
-    def updatePassword(self, userId, plainPassword):
+    def getAllMenuList(self):
+        pass
+
+    def getAllMenuAjax(self):
+        pass
+
+    def getMenus(self, menuIds):
         """
         Parameters:
-         - userId
-         - plainPassword
+         - menuIds
+        """
+        pass
+
+    def getAdminIdsByMenuId(self, menuId):
+        """
+        Parameters:
+         - menuId
+        """
+        pass
+
+    def getAllUserIds(self, menuId):
+        """
+        Parameters:
+         - menuId
         """
         pass
 
 
 class Client(Iface):
     """
-    AdminUser 서비스
+    AdminMenu 서비스
     """
     def __init__(self, iprot, oprot=None):
         self._iprot = self._oprot = iprot
@@ -80,199 +65,23 @@ class Client(Iface):
             self._oprot = oprot
         self._seqid = 0
 
-    def getAllAdminUserArray(self):
-        self.send_getAllAdminUserArray()
-        return self.recv_getAllAdminUserArray()
-
-    def send_getAllAdminUserArray(self):
-        self._oprot.writeMessageBegin('getAllAdminUserArray', TMessageType.CALL, self._seqid)
-        args = getAllAdminUserArray_args()
-        args.write(self._oprot)
-        self._oprot.writeMessageEnd()
-        self._oprot.trans.flush()
-
-    def recv_getAllAdminUserArray(self):
-        iprot = self._iprot
-        (fname, mtype, rseqid) = iprot.readMessageBegin()
-        if mtype == TMessageType.EXCEPTION:
-            x = TApplicationException()
-            x.read(iprot)
-            iprot.readMessageEnd()
-            raise x
-        result = getAllAdminUserArray_result()
-        result.read(iprot)
-        iprot.readMessageEnd()
-        if result.success is not None:
-            return result.success
-        if result.userException is not None:
-            raise result.userException
-        if result.systemException is not None:
-            raise result.systemException
-        raise TApplicationException(TApplicationException.MISSING_RESULT, "getAllAdminUserArray failed: unknown result")
-
-    def getUser(self, userId):
+    def getMenuList(self, isUse):
         """
         Parameters:
-         - userId
-        """
-        self.send_getUser(userId)
-        return self.recv_getUser()
-
-    def send_getUser(self, userId):
-        self._oprot.writeMessageBegin('getUser', TMessageType.CALL, self._seqid)
-        args = getUser_args()
-        args.userId = userId
-        args.write(self._oprot)
-        self._oprot.writeMessageEnd()
-        self._oprot.trans.flush()
-
-    def recv_getUser(self):
-        iprot = self._iprot
-        (fname, mtype, rseqid) = iprot.readMessageBegin()
-        if mtype == TMessageType.EXCEPTION:
-            x = TApplicationException()
-            x.read(iprot)
-            iprot.readMessageEnd()
-            raise x
-        result = getUser_result()
-        result.read(iprot)
-        iprot.readMessageEnd()
-        if result.success is not None:
-            return result.success
-        if result.userException is not None:
-            raise result.userException
-        if result.systemException is not None:
-            raise result.systemException
-        raise TApplicationException(TApplicationException.MISSING_RESULT, "getUser failed: unknown result")
-
-    def getAdminUserTag(self, userId):
-        """
-        Parameters:
-         - userId
-        """
-        self.send_getAdminUserTag(userId)
-        return self.recv_getAdminUserTag()
-
-    def send_getAdminUserTag(self, userId):
-        self._oprot.writeMessageBegin('getAdminUserTag', TMessageType.CALL, self._seqid)
-        args = getAdminUserTag_args()
-        args.userId = userId
-        args.write(self._oprot)
-        self._oprot.writeMessageEnd()
-        self._oprot.trans.flush()
-
-    def recv_getAdminUserTag(self):
-        iprot = self._iprot
-        (fname, mtype, rseqid) = iprot.readMessageBegin()
-        if mtype == TMessageType.EXCEPTION:
-            x = TApplicationException()
-            x.read(iprot)
-            iprot.readMessageEnd()
-            raise x
-        result = getAdminUserTag_result()
-        result.read(iprot)
-        iprot.readMessageEnd()
-        if result.success is not None:
-            return result.success
-        if result.userException is not None:
-            raise result.userException
-        if result.systemException is not None:
-            raise result.systemException
-        raise TApplicationException(TApplicationException.MISSING_RESULT, "getAdminUserTag failed: unknown result")
-
-    def getAdminUserMenu(self, userId):
-        """
-        Parameters:
-         - userId
-        """
-        self.send_getAdminUserMenu(userId)
-        return self.recv_getAdminUserMenu()
-
-    def send_getAdminUserMenu(self, userId):
-        self._oprot.writeMessageBegin('getAdminUserMenu', TMessageType.CALL, self._seqid)
-        args = getAdminUserMenu_args()
-        args.userId = userId
-        args.write(self._oprot)
-        self._oprot.writeMessageEnd()
-        self._oprot.trans.flush()
-
-    def recv_getAdminUserMenu(self):
-        iprot = self._iprot
-        (fname, mtype, rseqid) = iprot.readMessageBegin()
-        if mtype == TMessageType.EXCEPTION:
-            x = TApplicationException()
-            x.read(iprot)
-            iprot.readMessageEnd()
-            raise x
-        result = getAdminUserMenu_result()
-        result.read(iprot)
-        iprot.readMessageEnd()
-        if result.success is not None:
-            return result.success
-        if result.userException is not None:
-            raise result.userException
-        if result.systemException is not None:
-            raise result.systemException
-        raise TApplicationException(TApplicationException.MISSING_RESULT, "getAdminUserMenu failed: unknown result")
-
-    def getAllMenuIds(self, userId):
-        """
-        Parameters:
-         - userId
-        """
-        self.send_getAllMenuIds(userId)
-        return self.recv_getAllMenuIds()
-
-    def send_getAllMenuIds(self, userId):
-        self._oprot.writeMessageBegin('getAllMenuIds', TMessageType.CALL, self._seqid)
-        args = getAllMenuIds_args()
-        args.userId = userId
-        args.write(self._oprot)
-        self._oprot.writeMessageEnd()
-        self._oprot.trans.flush()
-
-    def recv_getAllMenuIds(self):
-        iprot = self._iprot
-        (fname, mtype, rseqid) = iprot.readMessageBegin()
-        if mtype == TMessageType.EXCEPTION:
-            x = TApplicationException()
-            x.read(iprot)
-            iprot.readMessageEnd()
-            raise x
-        result = getAllMenuIds_result()
-        result.read(iprot)
-        iprot.readMessageEnd()
-        if result.success is not None:
-            return result.success
-        if result.userException is not None:
-            raise result.userException
-        if result.systemException is not None:
-            raise result.systemException
-        raise TApplicationException(TApplicationException.MISSING_RESULT, "getAllMenuIds failed: unknown result")
-
-    def updateMyInfo(self, name, team, isUse, passwd):
-        """
-        Parameters:
-         - name
-         - team
          - isUse
-         - passwd
         """
-        self.send_updateMyInfo(name, team, isUse, passwd)
-        return self.recv_updateMyInfo()
+        self.send_getMenuList(isUse)
+        return self.recv_getMenuList()
 
-    def send_updateMyInfo(self, name, team, isUse, passwd):
-        self._oprot.writeMessageBegin('updateMyInfo', TMessageType.CALL, self._seqid)
-        args = updateMyInfo_args()
-        args.name = name
-        args.team = team
+    def send_getMenuList(self, isUse):
+        self._oprot.writeMessageBegin('getMenuList', TMessageType.CALL, self._seqid)
+        args = getMenuList_args()
         args.isUse = isUse
-        args.passwd = passwd
         args.write(self._oprot)
         self._oprot.writeMessageEnd()
         self._oprot.trans.flush()
 
-    def recv_updateMyInfo(self):
+    def recv_getMenuList(self):
         iprot = self._iprot
         (fname, mtype, rseqid) = iprot.readMessageBegin()
         if mtype == TMessageType.EXCEPTION:
@@ -280,7 +89,7 @@ class Client(Iface):
             x.read(iprot)
             iprot.readMessageEnd()
             raise x
-        result = updateMyInfo_result()
+        result = getMenuList_result()
         result.read(iprot)
         iprot.readMessageEnd()
         if result.success is not None:
@@ -289,27 +98,85 @@ class Client(Iface):
             raise result.userException
         if result.systemException is not None:
             raise result.systemException
-        raise TApplicationException(TApplicationException.MISSING_RESULT, "updateMyInfo failed: unknown result")
+        raise TApplicationException(TApplicationException.MISSING_RESULT, "getMenuList failed: unknown result")
 
-    def updatePassword(self, userId, plainPassword):
+    def getAllMenuList(self):
+        self.send_getAllMenuList()
+        return self.recv_getAllMenuList()
+
+    def send_getAllMenuList(self):
+        self._oprot.writeMessageBegin('getAllMenuList', TMessageType.CALL, self._seqid)
+        args = getAllMenuList_args()
+        args.write(self._oprot)
+        self._oprot.writeMessageEnd()
+        self._oprot.trans.flush()
+
+    def recv_getAllMenuList(self):
+        iprot = self._iprot
+        (fname, mtype, rseqid) = iprot.readMessageBegin()
+        if mtype == TMessageType.EXCEPTION:
+            x = TApplicationException()
+            x.read(iprot)
+            iprot.readMessageEnd()
+            raise x
+        result = getAllMenuList_result()
+        result.read(iprot)
+        iprot.readMessageEnd()
+        if result.success is not None:
+            return result.success
+        if result.userException is not None:
+            raise result.userException
+        if result.systemException is not None:
+            raise result.systemException
+        raise TApplicationException(TApplicationException.MISSING_RESULT, "getAllMenuList failed: unknown result")
+
+    def getAllMenuAjax(self):
+        self.send_getAllMenuAjax()
+        return self.recv_getAllMenuAjax()
+
+    def send_getAllMenuAjax(self):
+        self._oprot.writeMessageBegin('getAllMenuAjax', TMessageType.CALL, self._seqid)
+        args = getAllMenuAjax_args()
+        args.write(self._oprot)
+        self._oprot.writeMessageEnd()
+        self._oprot.trans.flush()
+
+    def recv_getAllMenuAjax(self):
+        iprot = self._iprot
+        (fname, mtype, rseqid) = iprot.readMessageBegin()
+        if mtype == TMessageType.EXCEPTION:
+            x = TApplicationException()
+            x.read(iprot)
+            iprot.readMessageEnd()
+            raise x
+        result = getAllMenuAjax_result()
+        result.read(iprot)
+        iprot.readMessageEnd()
+        if result.success is not None:
+            return result.success
+        if result.userException is not None:
+            raise result.userException
+        if result.systemException is not None:
+            raise result.systemException
+        raise TApplicationException(TApplicationException.MISSING_RESULT, "getAllMenuAjax failed: unknown result")
+
+    def getMenus(self, menuIds):
         """
         Parameters:
-         - userId
-         - plainPassword
+         - menuIds
         """
-        self.send_updatePassword(userId, plainPassword)
-        return self.recv_updatePassword()
+        self.send_getMenus(menuIds)
+        return self.recv_getMenus()
 
-    def send_updatePassword(self, userId, plainPassword):
-        self._oprot.writeMessageBegin('updatePassword', TMessageType.CALL, self._seqid)
-        args = updatePassword_args()
-        args.userId = userId
-        args.plainPassword = plainPassword
+    def send_getMenus(self, menuIds):
+        self._oprot.writeMessageBegin('getMenus', TMessageType.CALL, self._seqid)
+        args = getMenus_args()
+        args.menuIds = menuIds
         args.write(self._oprot)
         self._oprot.writeMessageEnd()
         self._oprot.trans.flush()
 
-    def recv_updatePassword(self):
+    def recv_getMenus(self):
         iprot = self._iprot
         (fname, mtype, rseqid) = iprot.readMessageBegin()
         if mtype == TMessageType.EXCEPTION:
@@ -317,7 +184,7 @@ class Client(Iface):
             x.read(iprot)
             iprot.readMessageEnd()
             raise x
-        result = updatePassword_result()
+        result = getMenus_result()
         result.read(iprot)
         iprot.readMessageEnd()
         if result.success is not None:
@@ -326,20 +193,89 @@ class Client(Iface):
             raise result.userException
         if result.systemException is not None:
             raise result.systemException
-        raise TApplicationException(TApplicationException.MISSING_RESULT, "updatePassword failed: unknown result")
+        raise TApplicationException(TApplicationException.MISSING_RESULT, "getMenus failed: unknown result")
+
+    def getAdminIdsByMenuId(self, menuId):
+        """
+        Parameters:
+         - menuId
+        """
+        self.send_getAdminIdsByMenuId(menuId)
+        return self.recv_getAdminIdsByMenuId()
+
+    def send_getAdminIdsByMenuId(self, menuId):
+        self._oprot.writeMessageBegin('getAdminIdsByMenuId', TMessageType.CALL, self._seqid)
+        args = getAdminIdsByMenuId_args()
+        args.menuId = menuId
+        args.write(self._oprot)
+        self._oprot.writeMessageEnd()
+        self._oprot.trans.flush()
+
+    def recv_getAdminIdsByMenuId(self):
+        iprot = self._iprot
+        (fname, mtype, rseqid) = iprot.readMessageBegin()
+        if mtype == TMessageType.EXCEPTION:
+            x = TApplicationException()
+            x.read(iprot)
+            iprot.readMessageEnd()
+            raise x
+        result = getAdminIdsByMenuId_result()
+        result.read(iprot)
+        iprot.readMessageEnd()
+        if result.success is not None:
+            return result.success
+        if result.userException is not None:
+            raise result.userException
+        if result.systemException is not None:
+            raise result.systemException
+        raise TApplicationException(TApplicationException.MISSING_RESULT, "getAdminIdsByMenuId failed: unknown result")
+
+    def getAllUserIds(self, menuId):
+        """
+        Parameters:
+         - menuId
+        """
+        self.send_getAllUserIds(menuId)
+        return self.recv_getAllUserIds()
+
+    def send_getAllUserIds(self, menuId):
+        self._oprot.writeMessageBegin('getAllUserIds', TMessageType.CALL, self._seqid)
+        args = getAllUserIds_args()
+        args.menuId = menuId
+        args.write(self._oprot)
+        self._oprot.writeMessageEnd()
+        self._oprot.trans.flush()
+
+    def recv_getAllUserIds(self):
+        iprot = self._iprot
+        (fname, mtype, rseqid) = iprot.readMessageBegin()
+        if mtype == TMessageType.EXCEPTION:
+            x = TApplicationException()
+            x.read(iprot)
+            iprot.readMessageEnd()
+            raise x
+        result = getAllUserIds_result()
+        result.read(iprot)
+        iprot.readMessageEnd()
+        if result.success is not None:
+            return result.success
+        if result.userException is not None:
+            raise result.userException
+        if result.systemException is not None:
+            raise result.systemException
+        raise TApplicationException(TApplicationException.MISSING_RESULT, "getAllUserIds failed: unknown result")
 
 
 class Processor(Iface, TProcessor):
     def __init__(self, handler):
         self._handler = handler
         self._processMap = {}
-        self._processMap["getAllAdminUserArray"] = Processor.process_getAllAdminUserArray
-        self._processMap["getUser"] = Processor.process_getUser
-        self._processMap["getAdminUserTag"] = Processor.process_getAdminUserTag
-        self._processMap["getAdminUserMenu"] = Processor.process_getAdminUserMenu
-        self._processMap["getAllMenuIds"] = Processor.process_getAllMenuIds
-        self._processMap["updateMyInfo"] = Processor.process_updateMyInfo
-        self._processMap["updatePassword"] = Processor.process_updatePassword
+        self._processMap["getMenuList"] = Processor.process_getMenuList
+        self._processMap["getAllMenuList"] = Processor.process_getAllMenuList
+        self._processMap["getAllMenuAjax"] = Processor.process_getAllMenuAjax
+        self._processMap["getMenus"] = Processor.process_getMenus
+        self._processMap["getAdminIdsByMenuId"] = Processor.process_getAdminIdsByMenuId
+        self._processMap["getAllUserIds"] = Processor.process_getAllUserIds
 
     def process(self, iprot, oprot):
         (name, type, seqid) = iprot.readMessageBegin()
@@ -356,177 +292,152 @@ class Processor(Iface, TProcessor):
             self._processMap[name](self, seqid, iprot, oprot)
         return True
 
-    def process_getAllAdminUserArray(self, seqid, iprot, oprot):
-        args = getAllAdminUserArray_args()
+    def process_getMenuList(self, seqid, iprot, oprot):
+        args = getMenuList_args()
         args.read(iprot)
         iprot.readMessageEnd()
-        result = getAllAdminUserArray_result()
+        result = getMenuList_result()
         try:
-            result.success = self._handler.getAllAdminUserArray()
+            result.success = self._handler.getMenuList(args.isUse)
             msg_type = TMessageType.REPLY
         except (TTransport.TTransportException, KeyboardInterrupt, SystemExit):
             raise
-        except Errors.ttypes.UserException as userException:
+        except cmssdk.Errors.ttypes.UserException as userException:
             msg_type = TMessageType.REPLY
             result.userException = userException
-        except Errors.ttypes.SystemException as systemException:
+        except cmssdk.Errors.ttypes.SystemException as systemException:
             msg_type = TMessageType.REPLY
             result.systemException = systemException
         except Exception as ex:
             msg_type = TMessageType.EXCEPTION
             logging.exception(ex)
             result = TApplicationException(TApplicationException.INTERNAL_ERROR, 'Internal error')
-        oprot.writeMessageBegin("getAllAdminUserArray", msg_type, seqid)
+        oprot.writeMessageBegin("getMenuList", msg_type, seqid)
         result.write(oprot)
         oprot.writeMessageEnd()
         oprot.trans.flush()
 
-    def process_getUser(self, seqid, iprot, oprot):
-        args = getUser_args()
+    def process_getAllMenuList(self, seqid, iprot, oprot):
+        args = getAllMenuList_args()
         args.read(iprot)
         iprot.readMessageEnd()
-        result = getUser_result()
+        result = getAllMenuList_result()
         try:
-            result.success = self._handler.getUser(args.userId)
+            result.success = self._handler.getAllMenuList()
             msg_type = TMessageType.REPLY
         except (TTransport.TTransportException, KeyboardInterrupt, SystemExit):
             raise
-        except Errors.ttypes.UserException as userException:
+        except cmssdk.Errors.ttypes.UserException as userException:
             msg_type = TMessageType.REPLY
             result.userException = userException
-        except Errors.ttypes.SystemException as systemException:
+        except cmssdk.Errors.ttypes.SystemException as systemException:
             msg_type = TMessageType.REPLY
             result.systemException = systemException
         except Exception as ex:
             msg_type = TMessageType.EXCEPTION
             logging.exception(ex)
             result = TApplicationException(TApplicationException.INTERNAL_ERROR, 'Internal error')
-        oprot.writeMessageBegin("getUser", msg_type, seqid)
+        oprot.writeMessageBegin("getAllMenuList", msg_type, seqid)
         result.write(oprot)
         oprot.writeMessageEnd()
         oprot.trans.flush()
 
-    def process_getAdminUserTag(self, seqid, iprot, oprot):
-        args = getAdminUserTag_args()
+    def process_getAllMenuAjax(self, seqid, iprot, oprot):
+        args = getAllMenuAjax_args()
         args.read(iprot)
         iprot.readMessageEnd()
-        result = getAdminUserTag_result()
+        result = getAllMenuAjax_result()
         try:
-            result.success = self._handler.getAdminUserTag(args.userId)
+            result.success = self._handler.getAllMenuAjax()
             msg_type = TMessageType.REPLY
         except (TTransport.TTransportException, KeyboardInterrupt, SystemExit):
             raise
-        except Errors.ttypes.UserException as userException:
+        except cmssdk.Errors.ttypes.UserException as userException:
             msg_type = TMessageType.REPLY
             result.userException = userException
-        except Errors.ttypes.SystemException as systemException:
+        except cmssdk.Errors.ttypes.SystemException as systemException:
             msg_type = TMessageType.REPLY
             result.systemException = systemException
         except Exception as ex:
             msg_type = TMessageType.EXCEPTION
             logging.exception(ex)
             result = TApplicationException(TApplicationException.INTERNAL_ERROR, 'Internal error')
-        oprot.writeMessageBegin("getAdminUserTag", msg_type, seqid)
+        oprot.writeMessageBegin("getAllMenuAjax", msg_type, seqid)
         result.write(oprot)
         oprot.writeMessageEnd()
         oprot.trans.flush()
 
-    def process_getAdminUserMenu(self, seqid, iprot, oprot):
-        args = getAdminUserMenu_args()
+    def process_getMenus(self, seqid, iprot, oprot):
+        args = getMenus_args()
         args.read(iprot)
         iprot.readMessageEnd()
-        result = getAdminUserMenu_result()
+        result = getMenus_result()
         try:
-            result.success = self._handler.getAdminUserMenu(args.userId)
+            result.success = self._handler.getMenus(args.menuIds)
             msg_type = TMessageType.REPLY
         except (TTransport.TTransportException, KeyboardInterrupt, SystemExit):
             raise
-        except Errors.ttypes.UserException as userException:
+        except cmssdk.Errors.ttypes.UserException as userException:
             msg_type = TMessageType.REPLY
             result.userException = userException
-        except Errors.ttypes.SystemException as systemException:
+        except cmssdk.Errors.ttypes.SystemException as systemException:
             msg_type = TMessageType.REPLY
             result.systemException = systemException
         except Exception as ex:
             msg_type = TMessageType.EXCEPTION
             logging.exception(ex)
             result = TApplicationException(TApplicationException.INTERNAL_ERROR, 'Internal error')
-        oprot.writeMessageBegin("getAdminUserMenu", msg_type, seqid)
+        oprot.writeMessageBegin("getMenus", msg_type, seqid)
         result.write(oprot)
         oprot.writeMessageEnd()
         oprot.trans.flush()
 
-    def process_getAllMenuIds(self, seqid, iprot, oprot):
-        args = getAllMenuIds_args()
+    def process_getAdminIdsByMenuId(self, seqid, iprot, oprot):
+        args = getAdminIdsByMenuId_args()
         args.read(iprot)
         iprot.readMessageEnd()
-        result = getAllMenuIds_result()
+        result = getAdminIdsByMenuId_result()
         try:
-            result.success = self._handler.getAllMenuIds(args.userId)
+            result.success = self._handler.getAdminIdsByMenuId(args.menuId)
             msg_type = TMessageType.REPLY
         except (TTransport.TTransportException, KeyboardInterrupt, SystemExit):
             raise
-        except Errors.ttypes.UserException as userException:
+        except cmssdk.Errors.ttypes.UserException as userException:
             msg_type = TMessageType.REPLY
             result.userException = userException
-        except Errors.ttypes.SystemException as systemException:
+        except cmssdk.Errors.ttypes.SystemException as systemException:
             msg_type = TMessageType.REPLY
             result.systemException = systemException
         except Exception as ex:
             msg_type = TMessageType.EXCEPTION
             logging.exception(ex)
             result = TApplicationException(TApplicationException.INTERNAL_ERROR, 'Internal error')
-        oprot.writeMessageBegin("getAllMenuIds", msg_type, seqid)
+        oprot.writeMessageBegin("getAdminIdsByMenuId", msg_type, seqid)
         result.write(oprot)
         oprot.writeMessageEnd()
         oprot.trans.flush()
 
-    def process_updateMyInfo(self, seqid, iprot, oprot):
-        args = updateMyInfo_args()
+    def process_getAllUserIds(self, seqid, iprot, oprot):
+        args = getAllUserIds_args()
         args.read(iprot)
         iprot.readMessageEnd()
-        result = updateMyInfo_result()
+        result = getAllUserIds_result()
         try:
-            result.success = self._handler.updateMyInfo(args.name, args.team, args.isUse, args.passwd)
+            result.success = self._handler.getAllUserIds(args.menuId)
             msg_type = TMessageType.REPLY
         except (TTransport.TTransportException, KeyboardInterrupt, SystemExit):
             raise
-        except Errors.ttypes.UserException as userException:
+        except cmssdk.Errors.ttypes.UserException as userException:
             msg_type = TMessageType.REPLY
             result.userException = userException
-        except Errors.ttypes.SystemException as systemException:
+        except cmssdk.Errors.ttypes.SystemException as systemException:
             msg_type = TMessageType.REPLY
             result.systemException = systemException
         except Exception as ex:
             msg_type = TMessageType.EXCEPTION
             logging.exception(ex)
             result = TApplicationException(TApplicationException.INTERNAL_ERROR, 'Internal error')
-        oprot.writeMessageBegin("updateMyInfo", msg_type, seqid)
-        result.write(oprot)
-        oprot.writeMessageEnd()
-        oprot.trans.flush()
-
-    def process_updatePassword(self, seqid, iprot, oprot):
-        args = updatePassword_args()
-        args.read(iprot)
-        iprot.readMessageEnd()
-        result = updatePassword_result()
-        try:
-            result.success = self._handler.updatePassword(args.userId, args.plainPassword)
-            msg_type = TMessageType.REPLY
-        except (TTransport.TTransportException, KeyboardInterrupt, SystemExit):
-            raise
-        except Errors.ttypes.UserException as userException:
-            msg_type = TMessageType.REPLY
-            result.userException = userException
-        except Errors.ttypes.SystemException as systemException:
-            msg_type = TMessageType.REPLY
-            result.systemException = systemException
-        except Exception as ex:
-            msg_type = TMessageType.EXCEPTION
-            logging.exception(ex)
-            result = TApplicationException(TApplicationException.INTERNAL_ERROR, 'Internal error')
-        oprot.writeMessageBegin("updatePassword", msg_type, seqid)
+        oprot.writeMessageBegin("getAllUserIds", msg_type, seqid)
         result.write(oprot)
         oprot.writeMessageEnd()
         oprot.trans.flush()
@@ -534,10 +445,19 @@ class Processor(Iface, TProcessor):
 # HELPER FUNCTIONS AND STRUCTURES
 
 
-class getAllAdminUserArray_args(object):
+class getMenuList_args(object):
+    """
+    Attributes:
+     - isUse
+    """
 
     thrift_spec = (
+        None,  # 0
+        (1, TType.BOOL, 'isUse', None, None, ),  # 1
     )
+
+    def __init__(self, isUse=None,):
+        self.isUse = isUse
 
     def read(self, iprot):
         if iprot._fast_decode is not None and isinstance(iprot.trans, TTransport.CReadableTransport) and self.thrift_spec is not None:
@@ -548,6 +468,11 @@ class getAllAdminUserArray_args(object):
             (fname, ftype, fid) = iprot.readFieldBegin()
             if ftype == TType.STOP:
                 break
+            if fid == 1:
+                if ftype == TType.BOOL:
+                    self.isUse = iprot.readBool()
+                else:
+                    iprot.skip(ftype)
             else:
                 iprot.skip(ftype)
             iprot.readFieldEnd()
@@ -557,7 +482,11 @@ class getAllAdminUserArray_args(object):
         if oprot._fast_encode is not None and self.thrift_spec is not None:
             oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
             return
-        oprot.writeStructBegin('getAllAdminUserArray_args')
+        oprot.writeStructBegin('getMenuList_args')
+        if self.isUse is not None:
+            oprot.writeFieldBegin('isUse', TType.BOOL, 1)
+            oprot.writeBool(self.isUse)
+            oprot.writeFieldEnd()
         oprot.writeFieldStop()
         oprot.writeStructEnd()
 
@@ -576,7 +505,7 @@ class getAllAdminUserArray_args(object):
         return not (self == other)
 
 
-class getAllAdminUserArray_result(object):
+class getMenuList_result(object):
     """
     Attributes:
      - success
@@ -585,9 +514,9 @@ class getAllAdminUserArray_result(object):
     """
 
     thrift_spec = (
-        (0, TType.LIST, 'success', (TType.STRUCT, (AdminUser, AdminUser.thrift_spec), False), None, ),  # 0
-        (1, TType.STRUCT, 'userException', (Errors.ttypes.UserException, Errors.ttypes.UserException.thrift_spec), None, ),  # 1
-        (2, TType.STRUCT, 'systemException', (Errors.ttypes.SystemException, Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
+        (0, TType.LIST, 'success', (TType.STRUCT, (AdminMenu, AdminMenu.thrift_spec), False), None, ),  # 0
+        (1, TType.STRUCT, 'userException', (cmssdk.Errors.ttypes.UserException, cmssdk.Errors.ttypes.UserException.thrift_spec), None, ),  # 1
+        (2, TType.STRUCT, 'systemException', (cmssdk.Errors.ttypes.SystemException, cmssdk.Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
     )
 
     def __init__(self, success=None, userException=None, systemException=None,):
@@ -609,7 +538,7 @@ class getAllAdminUserArray_result(object):
                     self.success = []
                     (_etype3, _size0) = iprot.readListBegin()
                     for _i4 in range(_size0):
-                        _elem5 = AdminUser()
+                        _elem5 = AdminMenu()
                         _elem5.read(iprot)
                         self.success.append(_elem5)
                     iprot.readListEnd()
@@ -617,13 +546,13 @@ class getAllAdminUserArray_result(object):
                     iprot.skip(ftype)
             elif fid == 1:
                 if ftype == TType.STRUCT:
-                    self.userException = Errors.ttypes.UserException()
+                    self.userException = cmssdk.Errors.ttypes.UserException()
                     self.userException.read(iprot)
                 else:
                     iprot.skip(ftype)
             elif fid == 2:
                 if ftype == TType.STRUCT:
-                    self.systemException = Errors.ttypes.SystemException()
+                    self.systemException = cmssdk.Errors.ttypes.SystemException()
                     self.systemException.read(iprot)
                 else:
                     iprot.skip(ftype)
@@ -636,7 +565,7 @@ class getAllAdminUserArray_result(object):
         if oprot._fast_encode is not None and self.thrift_spec is not None:
             oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
             return
-        oprot.writeStructBegin('getAllAdminUserArray_result')
+        oprot.writeStructBegin('getMenuList_result')
         if self.success is not None:
             oprot.writeFieldBegin('success', TType.LIST, 0)
             oprot.writeListBegin(TType.STRUCT, len(self.success))
@@ -670,19 +599,10 @@ class getAllAdminUserArray_result(object):
         return not (self == other)
 
 
-class getUser_args(object):
-    """
-    Attributes:
-     - userId
-    """
+class getAllMenuList_args(object):
 
     thrift_spec = (
-        None,  # 0
-        (1, TType.STRING, 'userId', 'UTF8', None, ),  # 1
     )
-
-    def __init__(self, userId=None,):
-        self.userId = userId
 
     def read(self, iprot):
         if iprot._fast_decode is not None and isinstance(iprot.trans, TTransport.CReadableTransport) and self.thrift_spec is not None:
@@ -693,11 +613,6 @@ class getUser_args(object):
             (fname, ftype, fid) = iprot.readFieldBegin()
             if ftype == TType.STOP:
                 break
-            if fid == 1:
-                if ftype == TType.STRING:
-                    self.userId = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
-                else:
-                    iprot.skip(ftype)
             else:
                 iprot.skip(ftype)
             iprot.readFieldEnd()
@@ -707,11 +622,7 @@ class getUser_args(object):
         if oprot._fast_encode is not None and self.thrift_spec is not None:
             oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
             return
-        oprot.writeStructBegin('getUser_args')
-        if self.userId is not None:
-            oprot.writeFieldBegin('userId', TType.STRING, 1)
-            oprot.writeString(self.userId.encode('utf-8') if sys.version_info[0] == 2 else self.userId)
-            oprot.writeFieldEnd()
+        oprot.writeStructBegin('getAllMenuList_args')
         oprot.writeFieldStop()
         oprot.writeStructEnd()
 
@@ -730,7 +641,7 @@ class getUser_args(object):
         return not (self == other)
 
 
-class getUser_result(object):
+class getAllMenuList_result(object):
     """
     Attributes:
      - success
@@ -739,155 +650,9 @@ class getUser_result(object):
     """
 
     thrift_spec = (
-        (0, TType.STRUCT, 'success', (AdminUser, AdminUser.thrift_spec), None, ),  # 0
-        (1, TType.STRUCT, 'userException', (Errors.ttypes.UserException, Errors.ttypes.UserException.thrift_spec), None, ),  # 1
-        (2, TType.STRUCT, 'systemException', (Errors.ttypes.SystemException, Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
-    )
-
-    def __init__(self, success=None, userException=None, systemException=None,):
-        self.success = success
-        self.userException = userException
-        self.systemException = systemException
-
-    def read(self, iprot):
-        if iprot._fast_decode is not None and isinstance(iprot.trans, TTransport.CReadableTransport) and self.thrift_spec is not None:
-            iprot._fast_decode(self, iprot, (self.__class__, self.thrift_spec))
-            return
-        iprot.readStructBegin()
-        while True:
-            (fname, ftype, fid) = iprot.readFieldBegin()
-            if ftype == TType.STOP:
-                break
-            if fid == 0:
-                if ftype == TType.STRUCT:
-                    self.success = AdminUser()
-                    self.success.read(iprot)
-                else:
-                    iprot.skip(ftype)
-            elif fid == 1:
-                if ftype == TType.STRUCT:
-                    self.userException = Errors.ttypes.UserException()
-                    self.userException.read(iprot)
-                else:
-                    iprot.skip(ftype)
-            elif fid == 2:
-                if ftype == TType.STRUCT:
-                    self.systemException = Errors.ttypes.SystemException()
-                    self.systemException.read(iprot)
-                else:
-                    iprot.skip(ftype)
-            else:
-                iprot.skip(ftype)
-            iprot.readFieldEnd()
-        iprot.readStructEnd()
-
-    def write(self, oprot):
-        if oprot._fast_encode is not None and self.thrift_spec is not None:
-            oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
-            return
-        oprot.writeStructBegin('getUser_result')
-        if self.success is not None:
-            oprot.writeFieldBegin('success', TType.STRUCT, 0)
-            self.success.write(oprot)
-            oprot.writeFieldEnd()
-        if self.userException is not None:
-            oprot.writeFieldBegin('userException', TType.STRUCT, 1)
-            self.userException.write(oprot)
-            oprot.writeFieldEnd()
-        if self.systemException is not None:
-            oprot.writeFieldBegin('systemException', TType.STRUCT, 2)
-            self.systemException.write(oprot)
-            oprot.writeFieldEnd()
-        oprot.writeFieldStop()
-        oprot.writeStructEnd()
-
-    def validate(self):
-        return
-
-    def __repr__(self):
-        L = ['%s=%r' % (key, value)
-             for key, value in self.__dict__.items()]
-        return '%s(%s)' % (self.__class__.__name__, ', '.join(L))
-
-    def __eq__(self, other):
-        return isinstance(other, self.__class__) and self.__dict__ == other.__dict__
-
-    def __ne__(self, other):
-        return not (self == other)
-
-
-class getAdminUserTag_args(object):
-    """
-    Attributes:
-     - userId
-    """
-
-    thrift_spec = (
-        None,  # 0
-        (1, TType.STRING, 'userId', 'UTF8', None, ),  # 1
-    )
-
-    def __init__(self, userId=None,):
-        self.userId = userId
-
-    def read(self, iprot):
-        if iprot._fast_decode is not None and isinstance(iprot.trans, TTransport.CReadableTransport) and self.thrift_spec is not None:
-            iprot._fast_decode(self, iprot, (self.__class__, self.thrift_spec))
-            return
-        iprot.readStructBegin()
-        while True:
-            (fname, ftype, fid) = iprot.readFieldBegin()
-            if ftype == TType.STOP:
-                break
-            if fid == 1:
-                if ftype == TType.STRING:
-                    self.userId = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
-                else:
-                    iprot.skip(ftype)
-            else:
-                iprot.skip(ftype)
-            iprot.readFieldEnd()
-        iprot.readStructEnd()
-
-    def write(self, oprot):
-        if oprot._fast_encode is not None and self.thrift_spec is not None:
-            oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
-            return
-        oprot.writeStructBegin('getAdminUserTag_args')
-        if self.userId is not None:
-            oprot.writeFieldBegin('userId', TType.STRING, 1)
-            oprot.writeString(self.userId.encode('utf-8') if sys.version_info[0] == 2 else self.userId)
-            oprot.writeFieldEnd()
-        oprot.writeFieldStop()
-        oprot.writeStructEnd()
-
-    def validate(self):
-        return
-
-    def __repr__(self):
-        L = ['%s=%r' % (key, value)
-             for key, value in self.__dict__.items()]
-        return '%s(%s)' % (self.__class__.__name__, ', '.join(L))
-
-    def __eq__(self, other):
-        return isinstance(other, self.__class__) and self.__dict__ == other.__dict__
-
-    def __ne__(self, other):
-        return not (self == other)
-
-
-class getAdminUserTag_result(object):
-    """
-    Attributes:
-     - success
-     - userException
-     - systemException
-    """
-
-    thrift_spec = (
-        (0, TType.LIST, 'success', (TType.I32, None, False), None, ),  # 0
-        (1, TType.STRUCT, 'userException', (Errors.ttypes.UserException, Errors.ttypes.UserException.thrift_spec), None, ),  # 1
-        (2, TType.STRUCT, 'systemException', (Errors.ttypes.SystemException, Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
+        (0, TType.LIST, 'success', (TType.STRUCT, (AdminMenu, AdminMenu.thrift_spec), False), None, ),  # 0
+        (1, TType.STRUCT, 'userException', (cmssdk.Errors.ttypes.UserException, cmssdk.Errors.ttypes.UserException.thrift_spec), None, ),  # 1
+        (2, TType.STRUCT, 'systemException', (cmssdk.Errors.ttypes.SystemException, cmssdk.Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
     )
 
     def __init__(self, success=None, userException=None, systemException=None,):
@@ -909,20 +674,21 @@ class getAdminUserTag_result(object):
                     self.success = []
                     (_etype10, _size7) = iprot.readListBegin()
                     for _i11 in range(_size7):
-                        _elem12 = iprot.readI32()
+                        _elem12 = AdminMenu()
+                        _elem12.read(iprot)
                         self.success.append(_elem12)
                     iprot.readListEnd()
                 else:
                     iprot.skip(ftype)
             elif fid == 1:
                 if ftype == TType.STRUCT:
-                    self.userException = Errors.ttypes.UserException()
+                    self.userException = cmssdk.Errors.ttypes.UserException()
                     self.userException.read(iprot)
                 else:
                     iprot.skip(ftype)
             elif fid == 2:
                 if ftype == TType.STRUCT:
-                    self.systemException = Errors.ttypes.SystemException()
+                    self.systemException = cmssdk.Errors.ttypes.SystemException()
                     self.systemException.read(iprot)
                 else:
                     iprot.skip(ftype)
@@ -935,12 +701,12 @@ class getAdminUserTag_result(object):
         if oprot._fast_encode is not None and self.thrift_spec is not None:
             oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
             return
-        oprot.writeStructBegin('getAdminUserTag_result')
+        oprot.writeStructBegin('getAllMenuList_result')
         if self.success is not None:
             oprot.writeFieldBegin('success', TType.LIST, 0)
-            oprot.writeListBegin(TType.I32, len(self.success))
+            oprot.writeListBegin(TType.STRUCT, len(self.success))
             for iter13 in self.success:
-                oprot.writeI32(iter13)
+                iter13.write(oprot)
             oprot.writeListEnd()
             oprot.writeFieldEnd()
         if self.userException is not None:
@@ -969,19 +735,10 @@ class getAdminUserTag_result(object):
         return not (self == other)
 
 
-class getAdminUserMenu_args(object):
-    """
-    Attributes:
-     - userId
-    """
+class getAllMenuAjax_args(object):
 
     thrift_spec = (
-        None,  # 0
-        (1, TType.STRING, 'userId', 'UTF8', None, ),  # 1
     )
-
-    def __init__(self, userId=None,):
-        self.userId = userId
 
     def read(self, iprot):
         if iprot._fast_decode is not None and isinstance(iprot.trans, TTransport.CReadableTransport) and self.thrift_spec is not None:
@@ -992,11 +749,6 @@ class getAdminUserMenu_args(object):
             (fname, ftype, fid) = iprot.readFieldBegin()
             if ftype == TType.STOP:
                 break
-            if fid == 1:
-                if ftype == TType.STRING:
-                    self.userId = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
-                else:
-                    iprot.skip(ftype)
             else:
                 iprot.skip(ftype)
             iprot.readFieldEnd()
@@ -1006,11 +758,7 @@ class getAdminUserMenu_args(object):
         if oprot._fast_encode is not None and self.thrift_spec is not None:
             oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
             return
-        oprot.writeStructBegin('getAdminUserMenu_args')
-        if self.userId is not None:
-            oprot.writeFieldBegin('userId', TType.STRING, 1)
-            oprot.writeString(self.userId.encode('utf-8') if sys.version_info[0] == 2 else self.userId)
-            oprot.writeFieldEnd()
+        oprot.writeStructBegin('getAllMenuAjax_args')
         oprot.writeFieldStop()
         oprot.writeStructEnd()
 
@@ -1029,7 +777,7 @@ class getAdminUserMenu_args(object):
         return not (self == other)
 
 
-class getAdminUserMenu_result(object):
+class getAllMenuAjax_result(object):
     """
     Attributes:
      - success
@@ -1038,9 +786,9 @@ class getAdminUserMenu_result(object):
     """
 
     thrift_spec = (
-        (0, TType.LIST, 'success', (TType.I32, None, False), None, ),  # 0
-        (1, TType.STRUCT, 'userException', (Errors.ttypes.UserException, Errors.ttypes.UserException.thrift_spec), None, ),  # 1
-        (2, TType.STRUCT, 'systemException', (Errors.ttypes.SystemException, Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
+        (0, TType.LIST, 'success', (TType.STRUCT, (AdminMenuAjax, AdminMenuAjax.thrift_spec), False), None, ),  # 0
+        (1, TType.STRUCT, 'userException', (cmssdk.Errors.ttypes.UserException, cmssdk.Errors.ttypes.UserException.thrift_spec), None, ),  # 1
+        (2, TType.STRUCT, 'systemException', (cmssdk.Errors.ttypes.SystemException, cmssdk.Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
     )
 
     def __init__(self, success=None, userException=None, systemException=None,):
@@ -1062,20 +810,21 @@ class getAdminUserMenu_result(object):
                     self.success = []
                     (_etype17, _size14) = iprot.readListBegin()
                     for _i18 in range(_size14):
-                        _elem19 = iprot.readI32()
+                        _elem19 = AdminMenuAjax()
+                        _elem19.read(iprot)
                         self.success.append(_elem19)
                     iprot.readListEnd()
                 else:
                     iprot.skip(ftype)
             elif fid == 1:
                 if ftype == TType.STRUCT:
-                    self.userException = Errors.ttypes.UserException()
+                    self.userException = cmssdk.Errors.ttypes.UserException()
                     self.userException.read(iprot)
                 else:
                     iprot.skip(ftype)
             elif fid == 2:
                 if ftype == TType.STRUCT:
-                    self.systemException = Errors.ttypes.SystemException()
+                    self.systemException = cmssdk.Errors.ttypes.SystemException()
                     self.systemException.read(iprot)
                 else:
                     iprot.skip(ftype)
@@ -1088,12 +837,12 @@ class getAdminUserMenu_result(object):
         if oprot._fast_encode is not None and self.thrift_spec is not None:
             oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
             return
-        oprot.writeStructBegin('getAdminUserMenu_result')
+        oprot.writeStructBegin('getAllMenuAjax_result')
         if self.success is not None:
             oprot.writeFieldBegin('success', TType.LIST, 0)
-            oprot.writeListBegin(TType.I32, len(self.success))
+            oprot.writeListBegin(TType.STRUCT, len(self.success))
             for iter20 in self.success:
-                oprot.writeI32(iter20)
+                iter20.write(oprot)
             oprot.writeListEnd()
             oprot.writeFieldEnd()
         if self.userException is not None:
@@ -1122,19 +871,19 @@ class getAdminUserMenu_result(object):
         return not (self == other)
 
 
-class getAllMenuIds_args(object):
+class getMenus_args(object):
     """
     Attributes:
-     - userId
+     - menuIds
     """
 
     thrift_spec = (
         None,  # 0
-        (1, TType.STRING, 'userId', 'UTF8', None, ),  # 1
+        (1, TType.LIST, 'menuIds', (TType.I32, None, False), None, ),  # 1
     )
 
-    def __init__(self, userId=None,):
-        self.userId = userId
+    def __init__(self, menuIds=None,):
+        self.menuIds = menuIds
 
     def read(self, iprot):
         if iprot._fast_decode is not None and isinstance(iprot.trans, TTransport.CReadableTransport) and self.thrift_spec is not None:
@@ -1146,8 +895,13 @@ class getAllMenuIds_args(object):
             if ftype == TType.STOP:
                 break
             if fid == 1:
-                if ftype == TType.STRING:
-                    self.userId = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
+                if ftype == TType.LIST:
+                    self.menuIds = []
+                    (_etype24, _size21) = iprot.readListBegin()
+                    for _i25 in range(_size21):
+                        _elem26 = iprot.readI32()
+                        self.menuIds.append(_elem26)
+                    iprot.readListEnd()
                 else:
                     iprot.skip(ftype)
             else:
@@ -1159,10 +913,13 @@ class getAllMenuIds_args(object):
         if oprot._fast_encode is not None and self.thrift_spec is not None:
             oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
             return
-        oprot.writeStructBegin('getAllMenuIds_args')
-        if self.userId is not None:
-            oprot.writeFieldBegin('userId', TType.STRING, 1)
-            oprot.writeString(self.userId.encode('utf-8') if sys.version_info[0] == 2 else self.userId)
+        oprot.writeStructBegin('getMenus_args')
+        if self.menuIds is not None:
+            oprot.writeFieldBegin('menuIds', TType.LIST, 1)
+            oprot.writeListBegin(TType.I32, len(self.menuIds))
+            for iter27 in self.menuIds:
+                oprot.writeI32(iter27)
+            oprot.writeListEnd()
             oprot.writeFieldEnd()
         oprot.writeFieldStop()
         oprot.writeStructEnd()
@@ -1182,7 +939,7 @@ class getAllMenuIds_args(object):
         return not (self == other)
 
 
-class getAllMenuIds_result(object):
+class getMenus_result(object):
     """
     Attributes:
      - success
@@ -1191,9 +948,9 @@ class getAllMenuIds_result(object):
     """
 
     thrift_spec = (
-        (0, TType.LIST, 'success', (TType.I32, None, False), None, ),  # 0
-        (1, TType.STRUCT, 'userException', (Errors.ttypes.UserException, Errors.ttypes.UserException.thrift_spec), None, ),  # 1
-        (2, TType.STRUCT, 'systemException', (Errors.ttypes.SystemException, Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
+        (0, TType.LIST, 'success', (TType.STRUCT, (AdminMenu, AdminMenu.thrift_spec), False), None, ),  # 0
+        (1, TType.STRUCT, 'userException', (cmssdk.Errors.ttypes.UserException, cmssdk.Errors.ttypes.UserException.thrift_spec), None, ),  # 1
+        (2, TType.STRUCT, 'systemException', (cmssdk.Errors.ttypes.SystemException, cmssdk.Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
     )
 
     def __init__(self, success=None, userException=None, systemException=None,):
@@ -1213,22 +970,23 @@ class getAllMenuIds_result(object):
             if fid == 0:
                 if ftype == TType.LIST:
                     self.success = []
-                    (_etype24, _size21) = iprot.readListBegin()
-                    for _i25 in range(_size21):
-                        _elem26 = iprot.readI32()
-                        self.success.append(_elem26)
+                    (_etype31, _size28) = iprot.readListBegin()
+                    for _i32 in range(_size28):
+                        _elem33 = AdminMenu()
+                        _elem33.read(iprot)
+                        self.success.append(_elem33)
                     iprot.readListEnd()
                 else:
                     iprot.skip(ftype)
             elif fid == 1:
                 if ftype == TType.STRUCT:
-                    self.userException = Errors.ttypes.UserException()
+                    self.userException = cmssdk.Errors.ttypes.UserException()
                     self.userException.read(iprot)
                 else:
                     iprot.skip(ftype)
             elif fid == 2:
                 if ftype == TType.STRUCT:
-                    self.systemException = Errors.ttypes.SystemException()
+                    self.systemException = cmssdk.Errors.ttypes.SystemException()
                     self.systemException.read(iprot)
                 else:
                     iprot.skip(ftype)
@@ -1241,12 +999,12 @@ class getAllMenuIds_result(object):
         if oprot._fast_encode is not None and self.thrift_spec is not None:
             oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
             return
-        oprot.writeStructBegin('getAllMenuIds_result')
+        oprot.writeStructBegin('getMenus_result')
         if self.success is not None:
             oprot.writeFieldBegin('success', TType.LIST, 0)
-            oprot.writeListBegin(TType.I32, len(self.success))
-            for iter27 in self.success:
-                oprot.writeI32(iter27)
+            oprot.writeListBegin(TType.STRUCT, len(self.success))
+            for iter34 in self.success:
+                iter34.write(oprot)
             oprot.writeListEnd()
             oprot.writeFieldEnd()
         if self.userException is not None:
@@ -1275,28 +1033,19 @@ class getAllMenuIds_result(object):
         return not (self == other)
 
 
-class updateMyInfo_args(object):
+class getAdminIdsByMenuId_args(object):
     """
     Attributes:
-     - name
-     - team
-     - isUse
-     - passwd
+     - menuId
     """
 
     thrift_spec = (
         None,  # 0
-        (1, TType.STRING, 'name', 'UTF8', None, ),  # 1
-        (2, TType.STRING, 'team', 'UTF8', None, ),  # 2
-        (3, TType.BOOL, 'isUse', None, None, ),  # 3
-        (4, TType.STRING, 'passwd', 'UTF8', None, ),  # 4
+        (1, TType.I32, 'menuId', None, None, ),  # 1
     )
 
-    def __init__(self, name=None, team=None, isUse=None, passwd=None,):
-        self.name = name
-        self.team = team
-        self.isUse = isUse
-        self.passwd = passwd
+    def __init__(self, menuId=None,):
+        self.menuId = menuId
 
     def read(self, iprot):
         if iprot._fast_decode is not None and isinstance(iprot.trans, TTransport.CReadableTransport) and self.thrift_spec is not None:
@@ -1308,23 +1057,8 @@ class updateMyInfo_args(object):
             if ftype == TType.STOP:
                 break
             if fid == 1:
-                if ftype == TType.STRING:
-                    self.name = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
-                else:
-                    iprot.skip(ftype)
-            elif fid == 2:
-                if ftype == TType.STRING:
-                    self.team = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
-                else:
-                    iprot.skip(ftype)
-            elif fid == 3:
-                if ftype == TType.BOOL:
-                    self.isUse = iprot.readBool()
-                else:
-                    iprot.skip(ftype)
-            elif fid == 4:
-                if ftype == TType.STRING:
-                    self.passwd = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
+                if ftype == TType.I32:
+                    self.menuId = iprot.readI32()
                 else:
                     iprot.skip(ftype)
             else:
@@ -1336,22 +1070,10 @@ class updateMyInfo_args(object):
         if oprot._fast_encode is not None and self.thrift_spec is not None:
             oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
             return
-        oprot.writeStructBegin('updateMyInfo_args')
-        if self.name is not None:
-            oprot.writeFieldBegin('name', TType.STRING, 1)
-            oprot.writeString(self.name.encode('utf-8') if sys.version_info[0] == 2 else self.name)
-            oprot.writeFieldEnd()
-        if self.team is not None:
-            oprot.writeFieldBegin('team', TType.STRING, 2)
-            oprot.writeString(self.team.encode('utf-8') if sys.version_info[0] == 2 else self.team)
-            oprot.writeFieldEnd()
-        if self.isUse is not None:
-            oprot.writeFieldBegin('isUse', TType.BOOL, 3)
-            oprot.writeBool(self.isUse)
-            oprot.writeFieldEnd()
-        if self.passwd is not None:
-            oprot.writeFieldBegin('passwd', TType.STRING, 4)
-            oprot.writeString(self.passwd.encode('utf-8') if sys.version_info[0] == 2 else self.passwd)
+        oprot.writeStructBegin('getAdminIdsByMenuId_args')
+        if self.menuId is not None:
+            oprot.writeFieldBegin('menuId', TType.I32, 1)
+            oprot.writeI32(self.menuId)
             oprot.writeFieldEnd()
         oprot.writeFieldStop()
         oprot.writeStructEnd()
@@ -1371,7 +1093,7 @@ class updateMyInfo_args(object):
         return not (self == other)
 
 
-class updateMyInfo_result(object):
+class getAdminIdsByMenuId_result(object):
     """
     Attributes:
      - success
@@ -1380,9 +1102,9 @@ class updateMyInfo_result(object):
     """
 
     thrift_spec = (
-        (0, TType.BOOL, 'success', None, None, ),  # 0
-        (1, TType.STRUCT, 'userException', (Errors.ttypes.UserException, Errors.ttypes.UserException.thrift_spec), None, ),  # 1
-        (2, TType.STRUCT, 'systemException', (Errors.ttypes.SystemException, Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
+        (0, TType.LIST, 'success', (TType.STRING, 'UTF8', False), None, ),  # 0
+        (1, TType.STRUCT, 'userException', (cmssdk.Errors.ttypes.UserException, cmssdk.Errors.ttypes.UserException.thrift_spec), None, ),  # 1
+        (2, TType.STRUCT, 'systemException', (cmssdk.Errors.ttypes.SystemException, cmssdk.Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
     )
 
     def __init__(self, success=None, userException=None, systemException=None,):
@@ -1400,19 +1122,24 @@ class updateMyInfo_result(object):
             if ftype == TType.STOP:
                 break
             if fid == 0:
-                if ftype == TType.BOOL:
-                    self.success = iprot.readBool()
+                if ftype == TType.LIST:
+                    self.success = []
+                    (_etype38, _size35) = iprot.readListBegin()
+                    for _i39 in range(_size35):
+                        _elem40 = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
+                        self.success.append(_elem40)
+                    iprot.readListEnd()
                 else:
                     iprot.skip(ftype)
             elif fid == 1:
                 if ftype == TType.STRUCT:
-                    self.userException = Errors.ttypes.UserException()
+                    self.userException = cmssdk.Errors.ttypes.UserException()
                     self.userException.read(iprot)
                 else:
                     iprot.skip(ftype)
             elif fid == 2:
                 if ftype == TType.STRUCT:
-                    self.systemException = Errors.ttypes.SystemException()
+                    self.systemException = cmssdk.Errors.ttypes.SystemException()
                     self.systemException.read(iprot)
                 else:
                     iprot.skip(ftype)
@@ -1425,10 +1152,13 @@ class updateMyInfo_result(object):
         if oprot._fast_encode is not None and self.thrift_spec is not None:
             oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
             return
-        oprot.writeStructBegin('updateMyInfo_result')
+        oprot.writeStructBegin('getAdminIdsByMenuId_result')
         if self.success is not None:
-            oprot.writeFieldBegin('success', TType.BOOL, 0)
-            oprot.writeBool(self.success)
+            oprot.writeFieldBegin('success', TType.LIST, 0)
+            oprot.writeListBegin(TType.STRING, len(self.success))
+            for iter41 in self.success:
+                oprot.writeString(iter41.encode('utf-8') if sys.version_info[0] == 2 else iter41)
+            oprot.writeListEnd()
             oprot.writeFieldEnd()
         if self.userException is not None:
             oprot.writeFieldBegin('userException', TType.STRUCT, 1)
@@ -1456,22 +1186,19 @@ class updateMyInfo_result(object):
         return not (self == other)
 
 
-class updatePassword_args(object):
+class getAllUserIds_args(object):
     """
     Attributes:
-     - userId
-     - plainPassword
+     - menuId
     """
 
     thrift_spec = (
         None,  # 0
-        (1, TType.STRING, 'userId', 'UTF8', None, ),  # 1
-        (2, TType.STRING, 'plainPassword', 'UTF8', None, ),  # 2
+        (1, TType.I32, 'menuId', None, None, ),  # 1
     )
 
-    def __init__(self, userId=None, plainPassword=None,):
-        self.userId = userId
-        self.plainPassword = plainPassword
+    def __init__(self, menuId=None,):
+        self.menuId = menuId
 
     def read(self, iprot):
         if iprot._fast_decode is not None and isinstance(iprot.trans, TTransport.CReadableTransport) and self.thrift_spec is not None:
@@ -1483,13 +1210,8 @@ class updatePassword_args(object):
             if ftype == TType.STOP:
                 break
             if fid == 1:
-                if ftype == TType.STRING:
-                    self.userId = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
-                else:
-                    iprot.skip(ftype)
-            elif fid == 2:
-                if ftype == TType.STRING:
-                    self.plainPassword = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
+                if ftype == TType.I32:
+                    self.menuId = iprot.readI32()
                 else:
                     iprot.skip(ftype)
             else:
@@ -1501,14 +1223,10 @@ class updatePassword_args(object):
         if oprot._fast_encode is not None and self.thrift_spec is not None:
             oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
             return
-        oprot.writeStructBegin('updatePassword_args')
-        if self.userId is not None:
-            oprot.writeFieldBegin('userId', TType.STRING, 1)
-            oprot.writeString(self.userId.encode('utf-8') if sys.version_info[0] == 2 else self.userId)
-            oprot.writeFieldEnd()
-        if self.plainPassword is not None:
-            oprot.writeFieldBegin('plainPassword', TType.STRING, 2)
-            oprot.writeString(self.plainPassword.encode('utf-8') if sys.version_info[0] == 2 else self.plainPassword)
+        oprot.writeStructBegin('getAllUserIds_args')
+        if self.menuId is not None:
+            oprot.writeFieldBegin('menuId', TType.I32, 1)
+            oprot.writeI32(self.menuId)
             oprot.writeFieldEnd()
         oprot.writeFieldStop()
         oprot.writeStructEnd()
@@ -1528,7 +1246,7 @@ class updatePassword_args(object):
         return not (self == other)
 
 
-class updatePassword_result(object):
+class getAllUserIds_result(object):
     """
     Attributes:
      - success
@@ -1537,9 +1255,9 @@ class updatePassword_result(object):
     """
 
     thrift_spec = (
-        (0, TType.BOOL, 'success', None, None, ),  # 0
-        (1, TType.STRUCT, 'userException', (Errors.ttypes.UserException, Errors.ttypes.UserException.thrift_spec), None, ),  # 1
-        (2, TType.STRUCT, 'systemException', (Errors.ttypes.SystemException, Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
+        (0, TType.LIST, 'success', (TType.STRING, 'UTF8', False), None, ),  # 0
+        (1, TType.STRUCT, 'userException', (cmssdk.Errors.ttypes.UserException, cmssdk.Errors.ttypes.UserException.thrift_spec), None, ),  # 1
+        (2, TType.STRUCT, 'systemException', (cmssdk.Errors.ttypes.SystemException, cmssdk.Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
     )
 
     def __init__(self, success=None, userException=None, systemException=None,):
@@ -1557,19 +1275,24 @@ class updatePassword_result(object):
             if ftype == TType.STOP:
                 break
             if fid == 0:
-                if ftype == TType.BOOL:
-                    self.success = iprot.readBool()
+                if ftype == TType.LIST:
+                    self.success = []
+                    (_etype45, _size42) = iprot.readListBegin()
+                    for _i46 in range(_size42):
+                        _elem47 = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
+                        self.success.append(_elem47)
+                    iprot.readListEnd()
                 else:
                     iprot.skip(ftype)
             elif fid == 1:
                 if ftype == TType.STRUCT:
-                    self.userException = Errors.ttypes.UserException()
+                    self.userException = cmssdk.Errors.ttypes.UserException()
                     self.userException.read(iprot)
                 else:
                     iprot.skip(ftype)
             elif fid == 2:
                 if ftype == TType.STRUCT:
-                    self.systemException = Errors.ttypes.SystemException()
+                    self.systemException = cmssdk.Errors.ttypes.SystemException()
                     self.systemException.read(iprot)
                 else:
                     iprot.skip(ftype)
@@ -1582,10 +1305,13 @@ class updatePassword_result(object):
         if oprot._fast_encode is not None and self.thrift_spec is not None:
             oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
             return
-        oprot.writeStructBegin('updatePassword_result')
+        oprot.writeStructBegin('getAllUserIds_result')
         if self.success is not None:
-            oprot.writeFieldBegin('success', TType.BOOL, 0)
-            oprot.writeBool(self.success)
+            oprot.writeFieldBegin('success', TType.LIST, 0)
+            oprot.writeListBegin(TType.STRING, len(self.success))
+            for iter48 in self.success:
+                oprot.writeString(iter48.encode('utf-8') if sys.version_info[0] == 2 else iter48)
+            oprot.writeListEnd()
             oprot.writeFieldEnd()
         if self.userException is not None:
             oprot.writeFieldBegin('userException', TType.STRUCT, 1)
