+++ lib/python/cmssdk/AdminMenu/constants.py
