+++ lib/python/cmssdk/AdminMenu/ttypes.py
@@ -10,7 +10,7 @@
 from thrift.Thrift import TType, TMessageType, TFrozenDict, TException, TApplicationException
 from thrift.protocol.TProtocol import TProtocolException
 import sys
-import Errors.ttypes
+import cmssdk.Errors.ttypes
 
 from thrift.transport import TTransport
 
@@ -154,6 +154,90 @@ class AdminMenu(object):
         if self.is_newtab is not None:
             oprot.writeFieldBegin('is_newtab', TType.BOOL, 9)
             oprot.writeBool(self.is_newtab)
+            oprot.writeFieldEnd()
+        oprot.writeFieldStop()
+        oprot.writeStructEnd()
+
+    def validate(self):
+        return
+
+    def __repr__(self):
+        L = ['%s=%r' % (key, value)
+             for key, value in self.__dict__.items()]
+        return '%s(%s)' % (self.__class__.__name__, ', '.join(L))
+
+    def __eq__(self, other):
+        return isinstance(other, self.__class__) and self.__dict__ == other.__dict__
+
+    def __ne__(self, other):
+        return not (self == other)
+
+
+class AdminMenuAjax(object):
+    """
+    Attributes:
+     - id
+     - menu_id
+     - ajax_url
+    """
+
+    thrift_spec = (
+        None,  # 0
+        (1, TType.I32, 'id', None, None, ),  # 1
+        (2, TType.I32, 'menu_id', None, None, ),  # 2
+        (3, TType.STRING, 'ajax_url', 'UTF8', None, ),  # 3
+    )
+
+    def __init__(self, id=None, menu_id=None, ajax_url=None,):
+        self.id = id
+        self.menu_id = menu_id
+        self.ajax_url = ajax_url
+
+    def read(self, iprot):
+        if iprot._fast_decode is not None and isinstance(iprot.trans, TTransport.CReadableTransport) and self.thrift_spec is not None:
+            iprot._fast_decode(self, iprot, (self.__class__, self.thrift_spec))
+            return
+        iprot.readStructBegin()
+        while True:
+            (fname, ftype, fid) = iprot.readFieldBegin()
+            if ftype == TType.STOP:
+                break
+            if fid == 1:
+                if ftype == TType.I32:
+                    self.id = iprot.readI32()
+                else:
+                    iprot.skip(ftype)
+            elif fid == 2:
+                if ftype == TType.I32:
+                    self.menu_id = iprot.readI32()
+                else:
+                    iprot.skip(ftype)
+            elif fid == 3:
+                if ftype == TType.STRING:
+                    self.ajax_url = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
+                else:
+                    iprot.skip(ftype)
+            else:
+                iprot.skip(ftype)
+            iprot.readFieldEnd()
+        iprot.readStructEnd()
+
+    def write(self, oprot):
+        if oprot._fast_encode is not None and self.thrift_spec is not None:
+            oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
+            return
+        oprot.writeStructBegin('AdminMenuAjax')
+        if self.id is not None:
+            oprot.writeFieldBegin('id', TType.I32, 1)
+            oprot.writeI32(self.id)
+            oprot.writeFieldEnd()
+        if self.menu_id is not None:
+            oprot.writeFieldBegin('menu_id', TType.I32, 2)
+            oprot.writeI32(self.menu_id)
+            oprot.writeFieldEnd()
+        if self.ajax_url is not None:
+            oprot.writeFieldBegin('ajax_url', TType.STRING, 3)
+            oprot.writeString(self.ajax_url.encode('utf-8') if sys.version_info[0] == 2 else self.ajax_url)
             oprot.writeFieldEnd()
         oprot.writeFieldStop()
         oprot.writeStructEnd()
