+++ lib/python/cmssdk/AdminTag/AdminTagService.py
@@ -17,65 +17,53 @@ from thrift.transport import TTransport
 
 
 class Iface(object):
-    """
-    AdminAuth 서비스
-    """
-    def hasHashAuth(self, hash, checkUrl, adminId):
+    def getAdminIdsFromTags(self, tag_ids):
         """
         Parameters:
-         - hash
-         - checkUrl
-         - adminId
+         - tag_ids
         """
         pass
 
-    def getCurrentHashArray(self, checkUrl, adminId):
+    def getAdminTagMenus(self, tag_id):
         """
         Parameters:
-         - checkUrl
-         - adminId
+         - tag_id
         """
         pass
 
-    def getAdminMenu(self, adminId):
+    def getMappedAdminMenuHashes(self, check_url, tag_id):
         """
         Parameters:
-         - adminId
+         - check_url
+         - tag_id
         """
         pass
 
 
 class Client(Iface):
-    """
-    AdminAuth 서비스
-    """
     def __init__(self, iprot, oprot=None):
         self._iprot = self._oprot = iprot
         if oprot is not None:
             self._oprot = oprot
         self._seqid = 0
 
-    def hasHashAuth(self, hash, checkUrl, adminId):
+    def getAdminIdsFromTags(self, tag_ids):
         """
         Parameters:
-         - hash
-         - checkUrl
-         - adminId
+         - tag_ids
         """
-        self.send_hasHashAuth(hash, checkUrl, adminId)
-        return self.recv_hasHashAuth()
+        self.send_getAdminIdsFromTags(tag_ids)
+        return self.recv_getAdminIdsFromTags()
 
-    def send_hasHashAuth(self, hash, checkUrl, adminId):
-        self._oprot.writeMessageBegin('hasHashAuth', TMessageType.CALL, self._seqid)
-        args = hasHashAuth_args()
-        args.hash = hash
-        args.checkUrl = checkUrl
-        args.adminId = adminId
+    def send_getAdminIdsFromTags(self, tag_ids):
+        self._oprot.writeMessageBegin('getAdminIdsFromTags', TMessageType.CALL, self._seqid)
+        args = getAdminIdsFromTags_args()
+        args.tag_ids = tag_ids
         args.write(self._oprot)
         self._oprot.writeMessageEnd()
         self._oprot.trans.flush()
 
-    def recv_hasHashAuth(self):
+    def recv_getAdminIdsFromTags(self):
         iprot = self._iprot
         (fname, mtype, rseqid) = iprot.readMessageBegin()
         if mtype == TMessageType.EXCEPTION:
@@ -83,7 +71,7 @@ class Client(Iface):
             x.read(iprot)
             iprot.readMessageEnd()
             raise x
-        result = hasHashAuth_result()
+        result = getAdminIdsFromTags_result()
         result.read(iprot)
         iprot.readMessageEnd()
         if result.success is not None:
@@ -92,27 +80,25 @@ class Client(Iface):
             raise result.userException
         if result.systemException is not None:
             raise result.systemException
-        raise TApplicationException(TApplicationException.MISSING_RESULT, "hasHashAuth failed: unknown result")
+        raise TApplicationException(TApplicationException.MISSING_RESULT, "getAdminIdsFromTags failed: unknown result")
 
-    def getCurrentHashArray(self, checkUrl, adminId):
+    def getAdminTagMenus(self, tag_id):
         """
         Parameters:
-         - checkUrl
-         - adminId
+         - tag_id
         """
-        self.send_getCurrentHashArray(checkUrl, adminId)
-        return self.recv_getCurrentHashArray()
+        self.send_getAdminTagMenus(tag_id)
+        return self.recv_getAdminTagMenus()
 
-    def send_getCurrentHashArray(self, checkUrl, adminId):
-        self._oprot.writeMessageBegin('getCurrentHashArray', TMessageType.CALL, self._seqid)
-        args = getCurrentHashArray_args()
-        args.checkUrl = checkUrl
-        args.adminId = adminId
+    def send_getAdminTagMenus(self, tag_id):
+        self._oprot.writeMessageBegin('getAdminTagMenus', TMessageType.CALL, self._seqid)
+        args = getAdminTagMenus_args()
+        args.tag_id = tag_id
         args.write(self._oprot)
         self._oprot.writeMessageEnd()
         self._oprot.trans.flush()
 
-    def recv_getCurrentHashArray(self):
+    def recv_getAdminTagMenus(self):
         iprot = self._iprot
         (fname, mtype, rseqid) = iprot.readMessageBegin()
         if mtype == TMessageType.EXCEPTION:
@@ -120,7 +106,7 @@ class Client(Iface):
             x.read(iprot)
             iprot.readMessageEnd()
             raise x
-        result = getCurrentHashArray_result()
+        result = getAdminTagMenus_result()
         result.read(iprot)
         iprot.readMessageEnd()
         if result.success is not None:
@@ -129,25 +115,27 @@ class Client(Iface):
             raise result.userException
         if result.systemException is not None:
             raise result.systemException
-        raise TApplicationException(TApplicationException.MISSING_RESULT, "getCurrentHashArray failed: unknown result")
+        raise TApplicationException(TApplicationException.MISSING_RESULT, "getAdminTagMenus failed: unknown result")
 
-    def getAdminMenu(self, adminId):
+    def getMappedAdminMenuHashes(self, check_url, tag_id):
         """
         Parameters:
-         - adminId
+         - check_url
+         - tag_id
         """
-        self.send_getAdminMenu(adminId)
-        return self.recv_getAdminMenu()
+        self.send_getMappedAdminMenuHashes(check_url, tag_id)
+        return self.recv_getMappedAdminMenuHashes()
 
-    def send_getAdminMenu(self, adminId):
-        self._oprot.writeMessageBegin('getAdminMenu', TMessageType.CALL, self._seqid)
-        args = getAdminMenu_args()
-        args.adminId = adminId
+    def send_getMappedAdminMenuHashes(self, check_url, tag_id):
+        self._oprot.writeMessageBegin('getMappedAdminMenuHashes', TMessageType.CALL, self._seqid)
+        args = getMappedAdminMenuHashes_args()
+        args.check_url = check_url
+        args.tag_id = tag_id
         args.write(self._oprot)
         self._oprot.writeMessageEnd()
         self._oprot.trans.flush()
 
-    def recv_getAdminMenu(self):
+    def recv_getMappedAdminMenuHashes(self):
         iprot = self._iprot
         (fname, mtype, rseqid) = iprot.readMessageBegin()
         if mtype == TMessageType.EXCEPTION:
@@ -155,7 +143,7 @@ class Client(Iface):
             x.read(iprot)
             iprot.readMessageEnd()
             raise x
-        result = getAdminMenu_result()
+        result = getMappedAdminMenuHashes_result()
         result.read(iprot)
         iprot.readMessageEnd()
         if result.success is not None:
@@ -164,16 +152,16 @@ class Client(Iface):
             raise result.userException
         if result.systemException is not None:
             raise result.systemException
-        raise TApplicationException(TApplicationException.MISSING_RESULT, "getAdminMenu failed: unknown result")
+        raise TApplicationException(TApplicationException.MISSING_RESULT, "getMappedAdminMenuHashes failed: unknown result")
 
 
 class Processor(Iface, TProcessor):
     def __init__(self, handler):
         self._handler = handler
         self._processMap = {}
-        self._processMap["hasHashAuth"] = Processor.process_hasHashAuth
-        self._processMap["getCurrentHashArray"] = Processor.process_getCurrentHashArray
-        self._processMap["getAdminMenu"] = Processor.process_getAdminMenu
+        self._processMap["getAdminIdsFromTags"] = Processor.process_getAdminIdsFromTags
+        self._processMap["getAdminTagMenus"] = Processor.process_getAdminTagMenus
+        self._processMap["getMappedAdminMenuHashes"] = Processor.process_getMappedAdminMenuHashes
 
     def process(self, iprot, oprot):
         (name, type, seqid) = iprot.readMessageBegin()
@@ -190,77 +178,77 @@ class Processor(Iface, TProcessor):
             self._processMap[name](self, seqid, iprot, oprot)
         return True
 
-    def process_hasHashAuth(self, seqid, iprot, oprot):
-        args = hasHashAuth_args()
+    def process_getAdminIdsFromTags(self, seqid, iprot, oprot):
+        args = getAdminIdsFromTags_args()
         args.read(iprot)
         iprot.readMessageEnd()
-        result = hasHashAuth_result()
+        result = getAdminIdsFromTags_result()
         try:
-            result.success = self._handler.hasHashAuth(args.hash, args.checkUrl, args.adminId)
+            result.success = self._handler.getAdminIdsFromTags(args.tag_ids)
             msg_type = TMessageType.REPLY
         except (TTransport.TTransportException, KeyboardInterrupt, SystemExit):
             raise
-        except Errors.ttypes.UserException as userException:
+        except cmssdk.Errors.ttypes.UserException as userException:
             msg_type = TMessageType.REPLY
             result.userException = userException
-        except Errors.ttypes.SystemException as systemException:
+        except cmssdk.Errors.ttypes.SystemException as systemException:
             msg_type = TMessageType.REPLY
             result.systemException = systemException
         except Exception as ex:
             msg_type = TMessageType.EXCEPTION
             logging.exception(ex)
             result = TApplicationException(TApplicationException.INTERNAL_ERROR, 'Internal error')
-        oprot.writeMessageBegin("hasHashAuth", msg_type, seqid)
+        oprot.writeMessageBegin("getAdminIdsFromTags", msg_type, seqid)
         result.write(oprot)
         oprot.writeMessageEnd()
         oprot.trans.flush()
 
-    def process_getCurrentHashArray(self, seqid, iprot, oprot):
-        args = getCurrentHashArray_args()
+    def process_getAdminTagMenus(self, seqid, iprot, oprot):
+        args = getAdminTagMenus_args()
         args.read(iprot)
         iprot.readMessageEnd()
-        result = getCurrentHashArray_result()
+        result = getAdminTagMenus_result()
         try:
-            result.success = self._handler.getCurrentHashArray(args.checkUrl, args.adminId)
+            result.success = self._handler.getAdminTagMenus(args.tag_id)
             msg_type = TMessageType.REPLY
         except (TTransport.TTransportException, KeyboardInterrupt, SystemExit):
             raise
-        except Errors.ttypes.UserException as userException:
+        except cmssdk.Errors.ttypes.UserException as userException:
             msg_type = TMessageType.REPLY
             result.userException = userException
-        except Errors.ttypes.SystemException as systemException:
+        except cmssdk.Errors.ttypes.SystemException as systemException:
             msg_type = TMessageType.REPLY
             result.systemException = systemException
         except Exception as ex:
             msg_type = TMessageType.EXCEPTION
             logging.exception(ex)
             result = TApplicationException(TApplicationException.INTERNAL_ERROR, 'Internal error')
-        oprot.writeMessageBegin("getCurrentHashArray", msg_type, seqid)
+        oprot.writeMessageBegin("getAdminTagMenus", msg_type, seqid)
         result.write(oprot)
         oprot.writeMessageEnd()
         oprot.trans.flush()
 
-    def process_getAdminMenu(self, seqid, iprot, oprot):
-        args = getAdminMenu_args()
+    def process_getMappedAdminMenuHashes(self, seqid, iprot, oprot):
+        args = getMappedAdminMenuHashes_args()
         args.read(iprot)
         iprot.readMessageEnd()
-        result = getAdminMenu_result()
+        result = getMappedAdminMenuHashes_result()
         try:
-            result.success = self._handler.getAdminMenu(args.adminId)
+            result.success = self._handler.getMappedAdminMenuHashes(args.check_url, args.tag_id)
             msg_type = TMessageType.REPLY
         except (TTransport.TTransportException, KeyboardInterrupt, SystemExit):
             raise
-        except Errors.ttypes.UserException as userException:
+        except cmssdk.Errors.ttypes.UserException as userException:
             msg_type = TMessageType.REPLY
             result.userException = userException
-        except Errors.ttypes.SystemException as systemException:
+        except cmssdk.Errors.ttypes.SystemException as systemException:
             msg_type = TMessageType.REPLY
             result.systemException = systemException
         except Exception as ex:
             msg_type = TMessageType.EXCEPTION
             logging.exception(ex)
             result = TApplicationException(TApplicationException.INTERNAL_ERROR, 'Internal error')
-        oprot.writeMessageBegin("getAdminMenu", msg_type, seqid)
+        oprot.writeMessageBegin("getMappedAdminMenuHashes", msg_type, seqid)
         result.write(oprot)
         oprot.writeMessageEnd()
         oprot.trans.flush()
@@ -268,25 +256,19 @@ class Processor(Iface, TProcessor):
 # HELPER FUNCTIONS AND STRUCTURES
 
 
-class hasHashAuth_args(object):
+class getAdminIdsFromTags_args(object):
     """
     Attributes:
-     - hash
-     - checkUrl
-     - adminId
+     - tag_ids
     """
 
     thrift_spec = (
         None,  # 0
-        (1, TType.STRING, 'hash', 'UTF8', None, ),  # 1
-        (2, TType.STRING, 'checkUrl', 'UTF8', None, ),  # 2
-        (3, TType.STRING, 'adminId', 'UTF8', None, ),  # 3
+        (1, TType.LIST, 'tag_ids', (TType.I32, None, False), None, ),  # 1
     )
 
-    def __init__(self, hash=None, checkUrl=None, adminId=None,):
-        self.hash = hash
-        self.checkUrl = checkUrl
-        self.adminId = adminId
+    def __init__(self, tag_ids=None,):
+        self.tag_ids = tag_ids
 
     def read(self, iprot):
         if iprot._fast_decode is not None and isinstance(iprot.trans, TTransport.CReadableTransport) and self.thrift_spec is not None:
@@ -298,18 +280,13 @@ class hasHashAuth_args(object):
             if ftype == TType.STOP:
                 break
             if fid == 1:
-                if ftype == TType.STRING:
-                    self.hash = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
-                else:
-                    iprot.skip(ftype)
-            elif fid == 2:
-                if ftype == TType.STRING:
-                    self.checkUrl = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
-                else:
-                    iprot.skip(ftype)
-            elif fid == 3:
-                if ftype == TType.STRING:
-                    self.adminId = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
+                if ftype == TType.LIST:
+                    self.tag_ids = []
+                    (_etype3, _size0) = iprot.readListBegin()
+                    for _i4 in range(_size0):
+                        _elem5 = iprot.readI32()
+                        self.tag_ids.append(_elem5)
+                    iprot.readListEnd()
                 else:
                     iprot.skip(ftype)
             else:
@@ -321,18 +298,13 @@ class hasHashAuth_args(object):
         if oprot._fast_encode is not None and self.thrift_spec is not None:
             oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
             return
-        oprot.writeStructBegin('hasHashAuth_args')
-        if self.hash is not None:
-            oprot.writeFieldBegin('hash', TType.STRING, 1)
-            oprot.writeString(self.hash.encode('utf-8') if sys.version_info[0] == 2 else self.hash)
-            oprot.writeFieldEnd()
-        if self.checkUrl is not None:
-            oprot.writeFieldBegin('checkUrl', TType.STRING, 2)
-            oprot.writeString(self.checkUrl.encode('utf-8') if sys.version_info[0] == 2 else self.checkUrl)
-            oprot.writeFieldEnd()
-        if self.adminId is not None:
-            oprot.writeFieldBegin('adminId', TType.STRING, 3)
-            oprot.writeString(self.adminId.encode('utf-8') if sys.version_info[0] == 2 else self.adminId)
+        oprot.writeStructBegin('getAdminIdsFromTags_args')
+        if self.tag_ids is not None:
+            oprot.writeFieldBegin('tag_ids', TType.LIST, 1)
+            oprot.writeListBegin(TType.I32, len(self.tag_ids))
+            for iter6 in self.tag_ids:
+                oprot.writeI32(iter6)
+            oprot.writeListEnd()
             oprot.writeFieldEnd()
         oprot.writeFieldStop()
         oprot.writeStructEnd()
@@ -352,164 +324,7 @@ class hasHashAuth_args(object):
         return not (self == other)
 
 
-class hasHashAuth_result(object):
-    """
-    Attributes:
-     - success
-     - userException
-     - systemException
-    """
-
-    thrift_spec = (
-        (0, TType.BOOL, 'success', None, None, ),  # 0
-        (1, TType.STRUCT, 'userException', (Errors.ttypes.UserException, Errors.ttypes.UserException.thrift_spec), None, ),  # 1
-        (2, TType.STRUCT, 'systemException', (Errors.ttypes.SystemException, Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
-    )
-
-    def __init__(self, success=None, userException=None, systemException=None,):
-        self.success = success
-        self.userException = userException
-        self.systemException = systemException
-
-    def read(self, iprot):
-        if iprot._fast_decode is not None and isinstance(iprot.trans, TTransport.CReadableTransport) and self.thrift_spec is not None:
-            iprot._fast_decode(self, iprot, (self.__class__, self.thrift_spec))
-            return
-        iprot.readStructBegin()
-        while True:
-            (fname, ftype, fid) = iprot.readFieldBegin()
-            if ftype == TType.STOP:
-                break
-            if fid == 0:
-                if ftype == TType.BOOL:
-                    self.success = iprot.readBool()
-                else:
-                    iprot.skip(ftype)
-            elif fid == 1:
-                if ftype == TType.STRUCT:
-                    self.userException = Errors.ttypes.UserException()
-                    self.userException.read(iprot)
-                else:
-                    iprot.skip(ftype)
-            elif fid == 2:
-                if ftype == TType.STRUCT:
-                    self.systemException = Errors.ttypes.SystemException()
-                    self.systemException.read(iprot)
-                else:
-                    iprot.skip(ftype)
-            else:
-                iprot.skip(ftype)
-            iprot.readFieldEnd()
-        iprot.readStructEnd()
-
-    def write(self, oprot):
-        if oprot._fast_encode is not None and self.thrift_spec is not None:
-            oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
-            return
-        oprot.writeStructBegin('hasHashAuth_result')
-        if self.success is not None:
-            oprot.writeFieldBegin('success', TType.BOOL, 0)
-            oprot.writeBool(self.success)
-            oprot.writeFieldEnd()
-        if self.userException is not None:
-            oprot.writeFieldBegin('userException', TType.STRUCT, 1)
-            self.userException.write(oprot)
-            oprot.writeFieldEnd()
-        if self.systemException is not None:
-            oprot.writeFieldBegin('systemException', TType.STRUCT, 2)
-            self.systemException.write(oprot)
-            oprot.writeFieldEnd()
-        oprot.writeFieldStop()
-        oprot.writeStructEnd()
-
-    def validate(self):
-        return
-
-    def __repr__(self):
-        L = ['%s=%r' % (key, value)
-             for key, value in self.__dict__.items()]
-        return '%s(%s)' % (self.__class__.__name__, ', '.join(L))
-
-    def __eq__(self, other):
-        return isinstance(other, self.__class__) and self.__dict__ == other.__dict__
-
-    def __ne__(self, other):
-        return not (self == other)
-
-
-class getCurrentHashArray_args(object):
-    """
-    Attributes:
-     - checkUrl
-     - adminId
-    """
-
-    thrift_spec = (
-        None,  # 0
-        (1, TType.STRING, 'checkUrl', 'UTF8', None, ),  # 1
-        (2, TType.STRING, 'adminId', 'UTF8', None, ),  # 2
-    )
-
-    def __init__(self, checkUrl=None, adminId=None,):
-        self.checkUrl = checkUrl
-        self.adminId = adminId
-
-    def read(self, iprot):
-        if iprot._fast_decode is not None and isinstance(iprot.trans, TTransport.CReadableTransport) and self.thrift_spec is not None:
-            iprot._fast_decode(self, iprot, (self.__class__, self.thrift_spec))
-            return
-        iprot.readStructBegin()
-        while True:
-            (fname, ftype, fid) = iprot.readFieldBegin()
-            if ftype == TType.STOP:
-                break
-            if fid == 1:
-                if ftype == TType.STRING:
-                    self.checkUrl = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
-                else:
-                    iprot.skip(ftype)
-            elif fid == 2:
-                if ftype == TType.STRING:
-                    self.adminId = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
-                else:
-                    iprot.skip(ftype)
-            else:
-                iprot.skip(ftype)
-            iprot.readFieldEnd()
-        iprot.readStructEnd()
-
-    def write(self, oprot):
-        if oprot._fast_encode is not None and self.thrift_spec is not None:
-            oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
-            return
-        oprot.writeStructBegin('getCurrentHashArray_args')
-        if self.checkUrl is not None:
-            oprot.writeFieldBegin('checkUrl', TType.STRING, 1)
-            oprot.writeString(self.checkUrl.encode('utf-8') if sys.version_info[0] == 2 else self.checkUrl)
-            oprot.writeFieldEnd()
-        if self.adminId is not None:
-            oprot.writeFieldBegin('adminId', TType.STRING, 2)
-            oprot.writeString(self.adminId.encode('utf-8') if sys.version_info[0] == 2 else self.adminId)
-            oprot.writeFieldEnd()
-        oprot.writeFieldStop()
-        oprot.writeStructEnd()
-
-    def validate(self):
-        return
-
-    def __repr__(self):
-        L = ['%s=%r' % (key, value)
-             for key, value in self.__dict__.items()]
-        return '%s(%s)' % (self.__class__.__name__, ', '.join(L))
-
-    def __eq__(self, other):
-        return isinstance(other, self.__class__) and self.__dict__ == other.__dict__
-
-    def __ne__(self, other):
-        return not (self == other)
-
-
-class getCurrentHashArray_result(object):
+class getAdminIdsFromTags_result(object):
     """
     Attributes:
      - success
@@ -519,8 +334,8 @@ class getCurrentHashArray_result(object):
 
     thrift_spec = (
         (0, TType.LIST, 'success', (TType.STRING, 'UTF8', False), None, ),  # 0
-        (1, TType.STRUCT, 'userException', (Errors.ttypes.UserException, Errors.ttypes.UserException.thrift_spec), None, ),  # 1
-        (2, TType.STRUCT, 'systemException', (Errors.ttypes.SystemException, Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
+        (1, TType.STRUCT, 'userException', (cmssdk.Errors.ttypes.UserException, cmssdk.Errors.ttypes.UserException.thrift_spec), None, ),  # 1
+        (2, TType.STRUCT, 'systemException', (cmssdk.Errors.ttypes.SystemException, cmssdk.Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
     )
 
     def __init__(self, success=None, userException=None, systemException=None,):
@@ -540,22 +355,22 @@ class getCurrentHashArray_result(object):
             if fid == 0:
                 if ftype == TType.LIST:
                     self.success = []
-                    (_etype3, _size0) = iprot.readListBegin()
-                    for _i4 in range(_size0):
-                        _elem5 = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
-                        self.success.append(_elem5)
+                    (_etype10, _size7) = iprot.readListBegin()
+                    for _i11 in range(_size7):
+                        _elem12 = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
+                        self.success.append(_elem12)
                     iprot.readListEnd()
                 else:
                     iprot.skip(ftype)
             elif fid == 1:
                 if ftype == TType.STRUCT:
-                    self.userException = Errors.ttypes.UserException()
+                    self.userException = cmssdk.Errors.ttypes.UserException()
                     self.userException.read(iprot)
                 else:
                     iprot.skip(ftype)
             elif fid == 2:
                 if ftype == TType.STRUCT:
-                    self.systemException = Errors.ttypes.SystemException()
+                    self.systemException = cmssdk.Errors.ttypes.SystemException()
                     self.systemException.read(iprot)
                 else:
                     iprot.skip(ftype)
@@ -568,12 +383,12 @@ class getCurrentHashArray_result(object):
         if oprot._fast_encode is not None and self.thrift_spec is not None:
             oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
             return
-        oprot.writeStructBegin('getCurrentHashArray_result')
+        oprot.writeStructBegin('getAdminIdsFromTags_result')
         if self.success is not None:
             oprot.writeFieldBegin('success', TType.LIST, 0)
             oprot.writeListBegin(TType.STRING, len(self.success))
-            for iter6 in self.success:
-                oprot.writeString(iter6.encode('utf-8') if sys.version_info[0] == 2 else iter6)
+            for iter13 in self.success:
+                oprot.writeString(iter13.encode('utf-8') if sys.version_info[0] == 2 else iter13)
             oprot.writeListEnd()
             oprot.writeFieldEnd()
         if self.userException is not None:
@@ -602,19 +417,19 @@ class getCurrentHashArray_result(object):
         return not (self == other)
 
 
-class getAdminMenu_args(object):
+class getAdminTagMenus_args(object):
     """
     Attributes:
-     - adminId
+     - tag_id
     """
 
     thrift_spec = (
         None,  # 0
-        (1, TType.STRING, 'adminId', 'UTF8', None, ),  # 1
+        (1, TType.I32, 'tag_id', None, None, ),  # 1
     )
 
-    def __init__(self, adminId=None,):
-        self.adminId = adminId
+    def __init__(self, tag_id=None,):
+        self.tag_id = tag_id
 
     def read(self, iprot):
         if iprot._fast_decode is not None and isinstance(iprot.trans, TTransport.CReadableTransport) and self.thrift_spec is not None:
@@ -626,8 +441,8 @@ class getAdminMenu_args(object):
             if ftype == TType.STOP:
                 break
             if fid == 1:
-                if ftype == TType.STRING:
-                    self.adminId = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
+                if ftype == TType.I32:
+                    self.tag_id = iprot.readI32()
                 else:
                     iprot.skip(ftype)
             else:
@@ -639,10 +454,10 @@ class getAdminMenu_args(object):
         if oprot._fast_encode is not None and self.thrift_spec is not None:
             oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
             return
-        oprot.writeStructBegin('getAdminMenu_args')
-        if self.adminId is not None:
-            oprot.writeFieldBegin('adminId', TType.STRING, 1)
-            oprot.writeString(self.adminId.encode('utf-8') if sys.version_info[0] == 2 else self.adminId)
+        oprot.writeStructBegin('getAdminTagMenus_args')
+        if self.tag_id is not None:
+            oprot.writeFieldBegin('tag_id', TType.I32, 1)
+            oprot.writeI32(self.tag_id)
             oprot.writeFieldEnd()
         oprot.writeFieldStop()
         oprot.writeStructEnd()
@@ -662,7 +477,7 @@ class getAdminMenu_args(object):
         return not (self == other)
 
 
-class getAdminMenu_result(object):
+class getAdminTagMenus_result(object):
     """
     Attributes:
      - success
@@ -671,9 +486,9 @@ class getAdminMenu_result(object):
     """
 
     thrift_spec = (
-        (0, TType.LIST, 'success', (TType.STRUCT, (AdminMenu, AdminMenu.thrift_spec), False), None, ),  # 0
-        (1, TType.STRUCT, 'userException', (Errors.ttypes.UserException, Errors.ttypes.UserException.thrift_spec), None, ),  # 1
-        (2, TType.STRUCT, 'systemException', (Errors.ttypes.SystemException, Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
+        (0, TType.LIST, 'success', (TType.I32, None, False), None, ),  # 0
+        (1, TType.STRUCT, 'userException', (cmssdk.Errors.ttypes.UserException, cmssdk.Errors.ttypes.UserException.thrift_spec), None, ),  # 1
+        (2, TType.STRUCT, 'systemException', (cmssdk.Errors.ttypes.SystemException, cmssdk.Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
     )
 
     def __init__(self, success=None, userException=None, systemException=None,):
@@ -693,23 +508,22 @@ class getAdminMenu_result(object):
             if fid == 0:
                 if ftype == TType.LIST:
                     self.success = []
-                    (_etype10, _size7) = iprot.readListBegin()
-                    for _i11 in range(_size7):
-                        _elem12 = AdminMenu()
-                        _elem12.read(iprot)
-                        self.success.append(_elem12)
+                    (_etype17, _size14) = iprot.readListBegin()
+                    for _i18 in range(_size14):
+                        _elem19 = iprot.readI32()
+                        self.success.append(_elem19)
                     iprot.readListEnd()
                 else:
                     iprot.skip(ftype)
             elif fid == 1:
                 if ftype == TType.STRUCT:
-                    self.userException = Errors.ttypes.UserException()
+                    self.userException = cmssdk.Errors.ttypes.UserException()
                     self.userException.read(iprot)
                 else:
                     iprot.skip(ftype)
             elif fid == 2:
                 if ftype == TType.STRUCT:
-                    self.systemException = Errors.ttypes.SystemException()
+                    self.systemException = cmssdk.Errors.ttypes.SystemException()
                     self.systemException.read(iprot)
                 else:
                     iprot.skip(ftype)
@@ -722,12 +536,177 @@ class getAdminMenu_result(object):
         if oprot._fast_encode is not None and self.thrift_spec is not None:
             oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
             return
-        oprot.writeStructBegin('getAdminMenu_result')
+        oprot.writeStructBegin('getAdminTagMenus_result')
         if self.success is not None:
             oprot.writeFieldBegin('success', TType.LIST, 0)
-            oprot.writeListBegin(TType.STRUCT, len(self.success))
-            for iter13 in self.success:
-                iter13.write(oprot)
+            oprot.writeListBegin(TType.I32, len(self.success))
+            for iter20 in self.success:
+                oprot.writeI32(iter20)
+            oprot.writeListEnd()
+            oprot.writeFieldEnd()
+        if self.userException is not None:
+            oprot.writeFieldBegin('userException', TType.STRUCT, 1)
+            self.userException.write(oprot)
+            oprot.writeFieldEnd()
+        if self.systemException is not None:
+            oprot.writeFieldBegin('systemException', TType.STRUCT, 2)
+            self.systemException.write(oprot)
+            oprot.writeFieldEnd()
+        oprot.writeFieldStop()
+        oprot.writeStructEnd()
+
+    def validate(self):
+        return
+
+    def __repr__(self):
+        L = ['%s=%r' % (key, value)
+             for key, value in self.__dict__.items()]
+        return '%s(%s)' % (self.__class__.__name__, ', '.join(L))
+
+    def __eq__(self, other):
+        return isinstance(other, self.__class__) and self.__dict__ == other.__dict__
+
+    def __ne__(self, other):
+        return not (self == other)
+
+
+class getMappedAdminMenuHashes_args(object):
+    """
+    Attributes:
+     - check_url
+     - tag_id
+    """
+
+    thrift_spec = (
+        None,  # 0
+        (1, TType.STRING, 'check_url', 'UTF8', None, ),  # 1
+        (2, TType.I32, 'tag_id', None, None, ),  # 2
+    )
+
+    def __init__(self, check_url=None, tag_id=None,):
+        self.check_url = check_url
+        self.tag_id = tag_id
+
+    def read(self, iprot):
+        if iprot._fast_decode is not None and isinstance(iprot.trans, TTransport.CReadableTransport) and self.thrift_spec is not None:
+            iprot._fast_decode(self, iprot, (self.__class__, self.thrift_spec))
+            return
+        iprot.readStructBegin()
+        while True:
+            (fname, ftype, fid) = iprot.readFieldBegin()
+            if ftype == TType.STOP:
+                break
+            if fid == 1:
+                if ftype == TType.STRING:
+                    self.check_url = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
+                else:
+                    iprot.skip(ftype)
+            elif fid == 2:
+                if ftype == TType.I32:
+                    self.tag_id = iprot.readI32()
+                else:
+                    iprot.skip(ftype)
+            else:
+                iprot.skip(ftype)
+            iprot.readFieldEnd()
+        iprot.readStructEnd()
+
+    def write(self, oprot):
+        if oprot._fast_encode is not None and self.thrift_spec is not None:
+            oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
+            return
+        oprot.writeStructBegin('getMappedAdminMenuHashes_args')
+        if self.check_url is not None:
+            oprot.writeFieldBegin('check_url', TType.STRING, 1)
+            oprot.writeString(self.check_url.encode('utf-8') if sys.version_info[0] == 2 else self.check_url)
+            oprot.writeFieldEnd()
+        if self.tag_id is not None:
+            oprot.writeFieldBegin('tag_id', TType.I32, 2)
+            oprot.writeI32(self.tag_id)
+            oprot.writeFieldEnd()
+        oprot.writeFieldStop()
+        oprot.writeStructEnd()
+
+    def validate(self):
+        return
+
+    def __repr__(self):
+        L = ['%s=%r' % (key, value)
+             for key, value in self.__dict__.items()]
+        return '%s(%s)' % (self.__class__.__name__, ', '.join(L))
+
+    def __eq__(self, other):
+        return isinstance(other, self.__class__) and self.__dict__ == other.__dict__
+
+    def __ne__(self, other):
+        return not (self == other)
+
+
+class getMappedAdminMenuHashes_result(object):
+    """
+    Attributes:
+     - success
+     - userException
+     - systemException
+    """
+
+    thrift_spec = (
+        (0, TType.LIST, 'success', (TType.STRING, 'UTF8', False), None, ),  # 0
+        (1, TType.STRUCT, 'userException', (cmssdk.Errors.ttypes.UserException, cmssdk.Errors.ttypes.UserException.thrift_spec), None, ),  # 1
+        (2, TType.STRUCT, 'systemException', (cmssdk.Errors.ttypes.SystemException, cmssdk.Errors.ttypes.SystemException.thrift_spec), None, ),  # 2
+    )
+
+    def __init__(self, success=None, userException=None, systemException=None,):
+        self.success = success
+        self.userException = userException
+        self.systemException = systemException
+
+    def read(self, iprot):
+        if iprot._fast_decode is not None and isinstance(iprot.trans, TTransport.CReadableTransport) and self.thrift_spec is not None:
+            iprot._fast_decode(self, iprot, (self.__class__, self.thrift_spec))
+            return
+        iprot.readStructBegin()
+        while True:
+            (fname, ftype, fid) = iprot.readFieldBegin()
+            if ftype == TType.STOP:
+                break
+            if fid == 0:
+                if ftype == TType.LIST:
+                    self.success = []
+                    (_etype24, _size21) = iprot.readListBegin()
+                    for _i25 in range(_size21):
+                        _elem26 = iprot.readString().decode('utf-8') if sys.version_info[0] == 2 else iprot.readString()
+                        self.success.append(_elem26)
+                    iprot.readListEnd()
+                else:
+                    iprot.skip(ftype)
+            elif fid == 1:
+                if ftype == TType.STRUCT:
+                    self.userException = cmssdk.Errors.ttypes.UserException()
+                    self.userException.read(iprot)
+                else:
+                    iprot.skip(ftype)
+            elif fid == 2:
+                if ftype == TType.STRUCT:
+                    self.systemException = cmssdk.Errors.ttypes.SystemException()
+                    self.systemException.read(iprot)
+                else:
+                    iprot.skip(ftype)
+            else:
+                iprot.skip(ftype)
+            iprot.readFieldEnd()
+        iprot.readStructEnd()
+
+    def write(self, oprot):
+        if oprot._fast_encode is not None and self.thrift_spec is not None:
+            oprot.trans.write(oprot._fast_encode(self, (self.__class__, self.thrift_spec)))
+            return
+        oprot.writeStructBegin('getMappedAdminMenuHashes_result')
+        if self.success is not None:
+            oprot.writeFieldBegin('success', TType.LIST, 0)
+            oprot.writeListBegin(TType.STRING, len(self.success))
+            for iter27 in self.success:
+                oprot.writeString(iter27.encode('utf-8') if sys.version_info[0] == 2 else iter27)
             oprot.writeListEnd()
             oprot.writeFieldEnd()
         if self.userException is not None:
