+++ lib/python/cmssdk/AdminTag/constants.py
