+++ lib/python/cmssdk/AdminUser/AdminUserService-remote
@@ -17,20 +17,21 @@ else:
 from thrift.transport import TTransport, TSocket, TSSLSocket, THttpClient
 from thrift.protocol.TBinaryProtocol import TBinaryProtocol
 
-from AdminMenu import AdminMenuService
-from AdminMenu.ttypes import *
+from cmssdk.AdminUser import AdminUserService
+from cmssdk.AdminUser.ttypes import *
 
 if len(sys.argv) <= 1 or sys.argv[1] == '--help':
     print('')
     print('Usage: ' + sys.argv[0] + ' [-h host[:port]] [-u url] [-f[ramed]] [-s[sl]] [-novalidate] [-ca_certs certs] [-keyfile keyfile] [-certfile certfile] function [arg1 [arg2...]]')
     print('')
     print('Functions:')
-    print('  AdminMenuCollection getMenuList(bool isUse)')
-    print('  AdminMenuCollection getAllMenuList()')
-    print('  AdminMenuAjaxCollection getAllMenuAjax()')
-    print('  AdminMenuCollection getMenus( menuIds)')
-    print('   getAdminIdsByMenuId(i32 menuId)')
-    print('   getAllUserIds(i32 menuId)')
+    print('  AdminUserCollection getAllAdminUserArray()')
+    print('  AdminUser getUser(string userId)')
+    print('   getAdminUserTag(string userId)')
+    print('   getAdminUserMenu(string userId)')
+    print('   getAllMenuIds(string userId)')
+    print('  bool updateMyInfo(string name, string team, bool isUse, string passwd)')
+    print('  bool updatePassword(string userId, string plainPassword)')
     print('')
     sys.exit(0)
 
@@ -107,44 +108,50 @@ else:
     else:
         transport = TTransport.TBufferedTransport(socket)
 protocol = TBinaryProtocol(transport)
-client = AdminMenuService.Client(protocol)
+client = AdminUserService.Client(protocol)
 transport.open()
 
-if cmd == 'getMenuList':
-    if len(args) != 1:
-        print('getMenuList requires 1 args')
-        sys.exit(1)
-    pp.pprint(client.getMenuList(eval(args[0]),))
-
-elif cmd == 'getAllMenuList':
+if cmd == 'getAllAdminUserArray':
     if len(args) != 0:
-        print('getAllMenuList requires 0 args')
+        print('getAllAdminUserArray requires 0 args')
         sys.exit(1)
-    pp.pprint(client.getAllMenuList())
+    pp.pprint(client.getAllAdminUserArray())
 
-elif cmd == 'getAllMenuAjax':
-    if len(args) != 0:
-        print('getAllMenuAjax requires 0 args')
-        sys.exit(1)
-    pp.pprint(client.getAllMenuAjax())
-
-elif cmd == 'getMenus':
+elif cmd == 'getUser':
     if len(args) != 1:
-        print('getMenus requires 1 args')
+        print('getUser requires 1 args')
         sys.exit(1)
-    pp.pprint(client.getMenus(eval(args[0]),))
+    pp.pprint(client.getUser(args[0],))
 
-elif cmd == 'getAdminIdsByMenuId':
+elif cmd == 'getAdminUserTag':
     if len(args) != 1:
-        print('getAdminIdsByMenuId requires 1 args')
+        print('getAdminUserTag requires 1 args')
         sys.exit(1)
-    pp.pprint(client.getAdminIdsByMenuId(eval(args[0]),))
+    pp.pprint(client.getAdminUserTag(args[0],))
 
-elif cmd == 'getAllUserIds':
+elif cmd == 'getAdminUserMenu':
     if len(args) != 1:
-        print('getAllUserIds requires 1 args')
+        print('getAdminUserMenu requires 1 args')
         sys.exit(1)
-    pp.pprint(client.getAllUserIds(eval(args[0]),))
+    pp.pprint(client.getAdminUserMenu(args[0],))
+
+elif cmd == 'getAllMenuIds':
+    if len(args) != 1:
+        print('getAllMenuIds requires 1 args')
+        sys.exit(1)
+    pp.pprint(client.getAllMenuIds(args[0],))
+
+elif cmd == 'updateMyInfo':
+    if len(args) != 4:
+        print('updateMyInfo requires 4 args')
+        sys.exit(1)
+    pp.pprint(client.updateMyInfo(args[0], args[1], eval(args[2]), args[3],))
+
+elif cmd == 'updatePassword':
+    if len(args) != 2:
+        print('updatePassword requires 2 args')
+        sys.exit(1)
+    pp.pprint(client.updatePassword(args[0], args[1],))
 
 else:
     print('Unrecognized method %s' % cmd)
