+++ lib/python/cmssdk/AdminUser/constants.py
