+++ lib/python/cmssdk/Errors/__init__.py
