+++ lib/python/cmssdk/Errors/constants.py
