+++ lib/python/cmssdk/Errors/ttypes.py
