+++ lib/python/test/test_CmsClient.py
@@ -1,0 +1,5 @@
+import unittest
+from cmssdk.CmsClient import CmsClient
+
+client = CmsClient('http://localhost:8000')
+user = client.adminUser.getUser('admin')
