+++ lib/thrift-idl/AdminAuth.thrift
@@ -1,6 +1,7 @@
 include "Errors.thrift"
 
 namespace php Ridibooks.Cms.Thrift.AdminAuth
+namespace py cmssdk.AdminAuth
 
 /**
  * AdminMenu 엔티티
