+++ lib/thrift-idl/AdminUser.thrift
@@ -3,6 +3,7 @@ include "AdminTag.thrift"
 include "AdminMenu.thrift"
 
 namespace php Ridibooks.Cms.Thrift.AdminUser
+namespace py cmssdk.AdminUser
 
 /**
  * AdminUser 엔티티
